% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{add_quality_flags}
\alias{add_quality_flags}
\title{Add Quality Flags}
\usage{
add_quality_flags(object, control = control_quality(), overwrite = FALSE)
}
\arguments{
\item{object}{(\code{HermesData}) \cr input.}

\item{control}{(\code{list}) \cr list of settings used to perform the quality control procedure.}

\item{overwrite}{(\code{flag})\cr whether previous results need to be overwritten.}
}
\value{
A HermesData object with added quality flags.
}
\description{
Add Quality Flags
}
\examples{
object <- HermesData(summarized_experiment)
control <- control_quality()
result <- add_quality_flags(object, control)
result <- add_quality_flags(object, control, overwrite = TRUE)
              
}
