% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{adjust_ps}
\alias{adjust_ps}
\title{Adjust Propensity Scores}
\usage{
adjust_ps(ps_obj, caliper = 0.2, ...)
}
\arguments{
\item{ps_obj}{A hermes_ps object}

\item{caliper}{Caliper distance}

\item{...}{Additional arguments}
}
\description{
Adjust Propensity Scores
}
