% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_na}
\alias{all_na}
\title{Checks Whether All Missing}
\usage{
all_na(x)
}
\arguments{
\item{x}{(\code{vector})\cr vector to check.}
}
\value{
Corresponding flag.
}
\description{
Internal function to check whether a whole vector is \code{NA}.
}
