% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertthat.R
\name{assertions}
\alias{assertions}
\title{Additional Assertions for \code{assert_that}}
\description{
We provide additional assertion functions which can be used together with \code{\link[assertthat:assert_that]{assertthat::assert_that()}}.
}
