% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{assess_balance}
\alias{assess_balance}
\title{Assess Balance}
\usage{
assess_balance(ps_obj, ...)
}
\arguments{
\item{ps_obj}{A hermes_ps object}

\item{...}{Additional arguments}
}
\description{
Assess Balance
}
