% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{autoplot,AnyHermesData-method}
\alias{autoplot,AnyHermesData-method}
\title{Generate all plots in default setting}
\usage{
\S4method{autoplot}{AnyHermesData}(object)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}
}
\value{
The \code{ggplot} objects from \code{draw_libsize_hist()}, \code{draw_libsize_qq}, \code{draw_libsize_densities},
\code{draw_nonzero_boxplot()} and \code{draw_genes_barplot()} functions with default settings.
}
\description{
Generate all plots in default setting
}
\examples{
result <- HermesData(summarized_experiment)
autoplot(result)
}
