% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{autoplot,HermesDataDiffExpr-method}
\alias{autoplot,HermesDataDiffExpr-method}
\title{Volcano plot for Differential Expression Analysis}
\usage{
\S4method{autoplot}{HermesDataDiffExpr}(object, adj.P.Val.threshold = 0.05, logFC.threshold = 2.5)
}
\arguments{
\item{object}{(\code{HermesDataDiffExpr})\cr input.}
}
\value{
A \code{ggplot} object for volvano plot with default settings.
}
\description{
This generates volcano plot for HermesDataDiffExpr object at default setting.
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
  add_quality_flags() \%>\%
  filter()
colData(object)$SEX <- factor(colData(object)$SEX)
res1 <- diff_expression(object, group = "SEX", method = "voom")
autoplot(res1)
res2 <- diff_expression(object, group = "SEX", method = "deseq2")
autoplot(res2)
}
