% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{autoplot,AnyHermesData-method}
\alias{autoplot,AnyHermesData-method}
\alias{plot_all}
\title{All standard plots in default setting}
\usage{
\S4method{autoplot}{AnyHermesData}(object)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}
}
\value{
A list with the \code{ggplot} objects from \code{\link[=draw_libsize_hist]{draw_libsize_hist()}}, \code{\link[=draw_libsize_qq]{draw_libsize_qq()}}, \link{draw_libsize_densities},
\code{\link[=draw_nonzero_boxplot]{draw_nonzero_boxplot()}} and \code{\link[=draw_genes_barplot]{draw_genes_barplot()}} functions with default settings.
}
\description{
This generates all standard plots - histogram and q-q plot of library sizes, density plot of the (log) counts
distributions, boxplot of the number of number of non-zero expressed genes per sample, and a stacked barplot of low
expression genes by chromosome at default setting.
}
\examples{
result <- HermesData(summarized_experiment)
autoplot(result)

}
