% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrdBookmark.R
\name{bookmarkList}
\alias{bookmarkList}
\title{List Word Bookmarks}
\usage{
bookmarkList(wrd = NULL)
}
\arguments{
\item{wrd}{A Word COM object. If \code{NULL}, the current
active Word session is used.}
}
\value{
A data frame with the following columns:
\describe{
\item{id}{Internal Word bookmark ID.}
\item{name}{Bookmark name.}
\item{pagenr}{Page number of the bookmark.}
\item{type}{Bookmark type inferred from the bookmark name.}
}
}
\description{
Returns a data frame containing all bookmarks in the active
Word document.
}
\details{
Bookmarks are returned in the order in which they appear
in the document.
}
\examples{
\dontrun{
bookmarkList()
}

}
