% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\docType{class}
\name{calc_cor}
\alias{calc_cor}
\alias{HermesDataCor-class}
\alias{.HermesDataCor}
\alias{HermesDataCor}
\title{Correlation between Samples of HermesData}
\usage{
calc_cor(object, assay_name = "counts", method = "pearson")
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr object to calculate the correlation.}

\item{assay_name}{(\code{string})\cr the assay name where the counts are located in.}

\item{method}{(\code{string})\cr the correlation method, see \code{\link[stats:cor]{stats::cor()}} for details.}
}
\value{
A \link{HermesDataCor} object which is an extension of a \link{matrix} with
additional quality flags in the slot \code{flag_data}: This contains the
\code{TechnicalFailureFlag} and \code{LowDepthFlag} columns describing the original
input samples.
}
\description{
This calculates the correlation matrix between the sample vectors of counts from
a specified assay.
}
\examples{
object <- HermesData(summarized_experiment)
calc_cor(object)
calc_cor(object, method = "spearman")
              
}
