% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_proportion}
\alias{check_proportion}
\alias{assert_proportion}
\alias{test_proportion}
\alias{expect_proportion}
\title{Check for proportion}
\usage{
check_proportion(x, null.ok = FALSE)

assert_proportion(
  x,
  null.ok = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)

test_proportion(x, null.ok = FALSE)

expect_proportion(x, null.ok = FALSE, info = NULL, label = vname(x))
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Check whether \code{x} is a proportion.
}
\examples{
check_proportion(0.25)
}
\seealso{
\code{\link{assertions}} for more details.
}
