% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlSession.R
\name{closeXl}
\alias{closeXl}
\title{Close Excel Session}
\usage{
closeXl(save = FALSE)
}
\arguments{
\item{save}{logical should the save dialog be displayed?}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Closes the current Excel session.
}
