% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colMeanZscores}
\alias{colMeanZscores}
\title{Generate gene signature using mean Z-score}
\usage{
colMeanZscores(x, x_assay = "counts")
}
\arguments{
\item{x}{object (\code{HermedData})}

\item{x_assay}{(\code{string}) name of the assay to use}
}
\value{
A named \code{\link{vector}} containing the mean Z-score
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function returns the mean Z-score from an assay stored in an
object of class \code{\link[Hermes:HermesData]{Hermes::HermesData}}
}
\examples{
object <- HermesData(summarized_experiment) |>
  add_quality_flags() |>
  filter()

colMeanZscores(object,"counts")
}
