% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colPrinComp1}
\alias{colPrinComp1}
\title{Generate gene signature using PC1}
\usage{
colPrinComp1(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{(\code{matrix})\cr containing numeric data.}

\item{center}{(\code{logical})\cr whether the variables should be zero centered.}

\item{scale}{(\code{logical})\cr whether the variables should be scaled to have unit variance.}
}
\value{
(\code{vector})\cr containing the value of PC1.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function returns the PC1 from an assay stored as a \code{matrix}.
}
\examples{
object <- HermesData(summarized_experiment) |>
  add_quality_flags() |>
  filter() |>
  normalize() |>
  assay("counts")

colPrinComp1(object)
}
