% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{compile_trajectories}
\alias{compile_trajectories}
\title{Compile Trajectories}
\usage{
compile_trajectories(ps_obj)
}
\arguments{
\item{ps_obj}{A hermes_ps object}
}
\description{
Compile Trajectories
}
