% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\docType{class}
\name{connect_biomart}
\alias{connect_biomart}
\alias{ConnectionBiomart-class}
\alias{.ConnectionBiomart}
\alias{ConnectionBiomart}
\title{Connection to Biomart}
\usage{
connect_biomart(prefix = c("ENSG", "GeneID"))
}
\arguments{
\item{prefix}{(\code{string})\cr gene ID prefix.}
}
\value{
\code{\link{ConnectionBiomart}} object.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Creates a connection object of class \code{\link{ConnectionBiomart}} which contains
the \code{biomaRt} object of class \code{\link[biomaRt:Mart-class]{biomaRt::Mart}} and the prefix of the object
which is used downstream for the query.
}
\details{
This connects to the Ensembl data base of Biomart for human genes.
}
\examples{
\dontrun{
connection <- connect_biomart("ENSG")
}
}
