% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_igis}
\alias{connect_igis}
\title{Connection to IGIS}
\usage{
connect_igis(prefix = c("ENSG", "GeneID"))
}
\arguments{
\item{prefix}{(\code{string})\cr gene ID prefix.}
}
\value{
\code{\link[igis:Igis]{igis::Igis}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Based on the prefix of the gene IDs, either creates a connection object of class
\code{\link[igis:Igis]{igis::Igis}} for version 4.0 (\code{ENSG}) or 3.0 (\code{GeneID}) (for human species genes).
}
\examples{
\dontrun{
connection <- connect_igis("ENSG")
}
}
