% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cor,AnyHermesData,ANY-method}
\alias{cor,AnyHermesData,ANY-method}
\alias{cor}
\title{Correlation Function}
\usage{
\S4method{cor}{AnyHermesData,ANY}(x, y = "counts", use = NULL, method = "pearson")
}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object to calculate the correlation.}

\item{y}{(\code{string})\cr the assay name where the counts are located in x (AnyHermesData object).}

\item{use}{not used.}

\item{method}{(\code{string})\cr the correlation coefficient (or covariance) to be computed, either "pearson", "kendall", or "spearman".}
}
\value{
A \link{HermesDataCor} object with calculated correlations and QC flags (technical failure and low depth).
}
\description{
Correlation Function
}
\examples{
object <- HermesData(summarized_experiment)
cor(x = object)
cor(x = object, method = "spearman")
              
}
