% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{counts,AnyHermesData-method}
\alias{counts,AnyHermesData-method}
\alias{counts}
\title{Counts Accessor and Setter}
\usage{
\S4method{counts}{AnyHermesData}(object, ...)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr object to access the counts from.}

\item{...}{not used.}

\item{value}{(\code{matrix})\cr what should the counts assay be replaced with.}
}
\value{
The counts assay.
}
\description{
These methods access and set the counts assay in a \code{\link{AnyHermesData}} object.
}
\examples{
a <- hermes_data
result <- counts(a)
class(result)
head(result)
}
