% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\docType{class}
\name{diff_expression}
\alias{diff_expression}
\alias{HermesDataDiffExpr-class}
\alias{.HermesDataDiffExpr}
\title{Functions for Differential Gene Expression Analysis}
\usage{
diff_expression(object, group, method = c("limma_voom", "deseq2"))
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{group}{(\code{string})\cr color of the bars filling.}

\item{method}{(\code{string})\cr we support limma-voom (limma_voom) and DESeq/DESeq2 (deseq2) tools.}
}
\value{
A \code{HermesDataDiffExpr} object.
}
\description{
This creates a histogram of the library sizes of the \link{HermesDataDiffExpr} object.
}
\examples{
object <- HermesData(summarized_experiment) \%>\% add_quality_flags() \%>\% filter()
diff_expression(object, "SEX", method = "limma_voom")
diff_expression(object, "SEX", method = "deseq2")

}
