% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\docType{class}
\name{diff_expression,AnyHermesData-method}
\alias{diff_expression,AnyHermesData-method}
\alias{differential}
\alias{HermesDataDiffExpr-class}
\alias{.HermesDataDiffExpr}
\alias{HermesDataDiffExpr}
\title{Differential Expression Analysis}
\usage{
\S4method{diff_expression}{AnyHermesData}(object, group, method = c("limma_voom", "deseq2"))
}
\arguments{
\item{object}{(\code{HermesData})\cr input. Note that this function only accepts un-normalized counts assay for analysis.}

\item{group}{(\code{vector})\cr A character vector indicating a factor variable with 2 levels in \code{\link{AnyHermesData}}.}

\item{method}{(\code{vector})\cr A character vector indicating desired method for differential expression analysis, see
details.}
}
\value{
A \link{HermesDataDiffExp} object with \code{log2_fc} (estimate of the log2 fold change between the 2 levels of the
provided factor), \code{stat} (the test statistic - which depends on the method used), \code{p_val} (the raw p-value), and
\code{adj_p_val} (the adjusted p-value) values from differential expression analysis for each feature / gene in the
\code{\link{HermesData}} object.
}
\description{
This method performs differential analysis using a method of preference.
}
\details{
Possible differential expression analysis methods are:
\itemize{
\item \code{limma_voom}: performs the differential expression analysis with the \code{voom} method from the \code{limma} package
(via \code{\link[limma:voom]{limma::voom()}}, \code{\link[limma:lmFit]{limma::lmFit()}} and \code{\link[limma:ebayes]{limma::eBayes()}}) for given counts in a \link{AnyHermesData} object and
a corresponding \code{design} matrix.
\item \code{deseq2}: performs the differential expression analysis with \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}} for a given \link{AnyHermesData} input
and \code{design} matrix.
}
}
\note{
Also validate that the \code{\link{HermesDataDiffExpr}} object contains columnns: \code{log2_fc}, \code{stat}, \code{p_val}, and
\code{adj_p_val} (the adjusted p-value).
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
  add_quality_flags() \%>\%
  filter() \%>\%
colData(object)$SEX <- factor(colData(object)$SEX)
res1 <- diff_expression(object, "SEX", method = "limma_voom")
head(res1)
res2 <- diff_expression(object, "SEX", method = "deseq2")
head(res2)

}
\references{
\insertRef{limma_package}{hermes}

\insertRef{voom_method}{hermes}

\insertRef{DESeq2_package}{hermes}
}
