% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_genes_barplot}
\alias{draw_genes_barplot}
\title{Stacked Barplot of Low Expression Genes by Chromosome}
\usage{
draw_genes_barplot(
  object,
  chromosomes = c(1:22, "X", "Y", "MT"),
  include_others = TRUE
)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{chromosomes}{(\code{character})\cr names of the chromosomes which should be displayed.}

\item{include_others}{(\code{flag})\cr option to show the chromosomes not in \code{chromosomes} as "Others".}
}
\value{
The \code{ggplot} object with the histogram.
}
\description{
This creates a barplot of chromosomes for the \link{AnyHermesData} object with the proportions of low expression genes.
}
\examples{
object <- HermesData(summarized_experiment)

# Display chromosomes 1-22, X, Y, and MT. Other chromosomes are displayed in "Others".
draw_genes_barplot(object)

# Display chromosomes 1 and 2. Other chromosomes are displayed in "Others".
draw_genes_barplot(object, chromosomes = c("1", "2"))

# Display chromosomes 1 and 2 only.
draw_genes_barplot(object, chromosomes = c("1", "2"), include_others = FALSE)

}
