% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Heatmap for Gene Expression Counts}
\usage{
draw_heatmap(
  object,
  assay_name = "counts",
  color_extremes = c(0.01, 0.99),
  col_data_annotation = NULL,
  ...
)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{assay_name}{(\code{string})\cr selects assay from input.}

\item{color_extremes}{(\code{numeric})\cr percentiles of dataset intended to be displayed.}

\item{col_data_annotation}{(\code{string} or \code{NULL})\cr optional grouping variable, taken
from input sample variables.}

\item{...(`string`}{or \code{NULL} or \verb{??})\cr additional arguments to pass to the ComplexHeatmap function}
}
\value{
The \code{ComplexHeatmap::Heatmap} heatmap
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This produces a heatmap of the normalized gene expression counts and groups by various factors available in the assay dataset
}
\examples{

result <- hermes_data \%>\%
  normalize(methods = "voom") \%>\%
  add_quality_flags()

result2 <- hermes::filter(result, what = "genes")


draw_heatmap(
  object = result2,
  col_data_annotation = c("COUNTRY")
)

draw_heatmap(
  object = result2,
  assay_name = "counts",
  color_extremes = c(0.001, 0.999),
  col_data_annotation = c("AGEGRP")
)

draw_heatmap(
  object = result2,
  assay_name = "counts",
  color_extremes = c(0.25, 0.75),
  col_data_annotation = c("Ethnic")
)
}
