% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_libsize_hist}
\alias{draw_libsize_hist}
\title{Histogram of Library Sizes}
\usage{
draw_libsize_hist(object, bins = 30L, fill = "darkgrey")
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{bins}{(\code{count})\cr number of evenly distributed groups desired.}

\item{fill}{(\code{string})\cr color of the bars filling.}
}
\value{
The \code{ggplot} object with the histogram.
}
\description{
This creates a histogram of the library sizes of the \link{AnyHermesData} object.
}
\examples{
result <- HermesData(summarized_experiment)
draw_libsize_hist(result)
draw_libsize_hist(result, bins = 10L, fill = "blue")
}
