% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_libsize_qq}
\alias{draw_libsize_qq}
\title{Q-Q plot of Library Sizes.}
\usage{
draw_libsize_qq(object, colour = "grey", linetype = "dashed")
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{colour}{(\code{string})\cr color of QQ-line, default colour grey.}

\item{linetype}{(\code{string})\cr linetype of  QQ-line, default linetype dashed.}
}
\value{
The \code{ggplot} object with the histogram.
}
\description{
This creates a Q-Q plot of the library sizes of the \link{HermesData} object.
}
\examples{
result <- hermes:::.HermesData(summarized_experiment)
draw_libsize_qq(result)
draw_libsize_qq(result, colour = "blue", linetype = "solid")
}
