% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_nonzero_boxplot}
\alias{draw_nonzero_boxplot}
\title{Boxplot of Non-zero Genes}
\usage{
draw_nonzero_boxplot(object, jitter = 0.2, alpha = 1/4)
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{jitter}{(\code{number})\cr \code{geom_point} aesthetic parameter.}

\item{alpha}{(\code{number})\cr \code{geom_point} aesthetic parameter.}
}
\value{
The \code{ggplot} object with the histogram.
}
\description{
This draws a boxplot, with overlaid data points, of the number of
non-zero expressed genes per sample.
}
\examples{
result <- HermesData(summarized_experiment)
draw_nonzero_boxplot(result)
draw_nonzero_boxplot(result, jitter = 0.1, alpha = 1/3)

}
