% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_top_barplot}
\alias{draw_top_barplot}
\title{Barplot top genes}
\usage{
draw_top_barplot(
  object,
  ylab = "Averaged Counts",
  title = "Top most expressed genes"
)
}
\arguments{
\item{object}{(\code{HermesDataTopGenes})\cr input.}

\item{ylab}{(\code{string})\cr input.}

\item{title}{(\code{string})\cr input.}
}
\value{
The \code{ggplot} object with the histogram.
}
\description{
Create a bar plot where the y is continuos and heights of the bars represent
the values in the data, and x is discrete variable. The input is a the data frame.
the data frame can be created with the function top_genes().
}
\examples{
result <- HermesData(summarized_experiment)

object <- top_genes(object = result)
draw_top_barplot(object)

object <- top_genes(result, n_top = NULL, min_threshold = 50000)
draw_top_barplot(object)

object <- top_genes(result, summary_fun = rowMax)
draw_top_barplot(object, ylab = "Maximum Count")
 

}
