% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{filter,AnyHermesData-method}
\alias{filter,AnyHermesData-method}
\alias{filter}
\title{Filter HermesData on Subset Passing Default QC Flags}
\usage{
\S4method{filter}{AnyHermesData}(x)
}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object to filter.}
}
\value{
The filtered \link{AnyHermesData} object.
}
\description{
This is a short cut to subset a \link{AnyHermesData} object for the default QC flag,
i.e. only features without low expression (\code{LowExpressionFlag}) and without samples
without low depth (\code{LowDepthFlag}) or technical failure (\code{TechnicalFailureFlag})
remain in the returend subset.
}
\note{
Internal implementation cannot use the \code{\link[=subset]{subset()}} method since that
requires non-standard evaluation of arguments.
}
\examples{
a <- HermesData(summarized_experiment)
dim(a)
result <- filter(a)
dim(result)

}
