% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{fit_ps}
\alias{fit_ps}
\title{Fit Propensity Score Model}
\usage{
fit_ps(hcru_obj, ...)
}
\arguments{
\item{hcru_obj}{A hermes_hcru object}

\item{...}{Additional arguments}
}
\description{
Fit Propensity Score Model
}
