% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{flag_tech_failure}
\alias{flag_tech_failure}
\title{Flag Technical Failures}
\usage{
flag_tech_failure(object, sample_ids)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{sample_ids}{(\code{list}) \cr list of sample IDs to be flagged manually as technical failures.}
}
\value{
HermesData object with modified technical failure flags.
}
\description{
Setter function which allows the user to define a sample manually as a technical failure.
}
\examples{
# Manually flag technical failures in a HermesData object.
object <- HermesData(summarized_experiment)
result <- flag_tech_failure(object, "06520101B0017R")
result <- flag_tech_failure(object, c("06520101B0017R", "06520047C0017R"))

}
