% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSpec-class.R
\name{gene_spec}
\alias{gene_spec}
\title{GeneSpec Constructor}
\usage{
gene_spec(genes, fun = NULL, fun_name = deparse(substitute(fun)))
}
\arguments{
\item{genes}{(\code{character} or \code{NULL})\cr the gene IDs.}

\item{fun}{(\code{function} or \code{NULL})\cr summary function. If \code{NULL} is
used then multiple genes are not summarized but returned as a matrix from the
\code{extract} method.}

\item{fun_name}{(\code{string})\cr name of the summary function.}
}
\value{
A new \code{\link{GeneSpec}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates a new \code{\link{GeneSpec}} object.
}
