% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_codes}
\alias{get_codes}
\title{get phenotype codes}
\usage{
get_codes(pheno_id)
}
\arguments{
\item{pheno_id}{a string, a valid phenotype id (see \code{get_phenotypes()})}
}
\value{
a data.table
}
\description{
get phenotype codes
}
