% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_metadata}
\alias{get_metadata}
\title{get phenotype meta-data}
\usage{
get_metadata(pheno_id)
}
\arguments{
\item{pheno_id}{a string, a valid phenotype id (see \code{get_phenotypes()})}
}
\value{
a data.table
}
\description{
get phenotype meta-data
}
