% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{get_quality_flags}
\alias{get_quality_flags}
\alias{get_tech_failure}
\alias{get_low_depth}
\alias{get_low_expression}
\title{Get Quality Flags}
\usage{
get_tech_failure(object)

get_low_depth(object)

get_low_expression(object)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}
}
\value{
Named logical vector containing the failure flags for all samples or genes,
respectively.
}
\description{
Separate getter functions which allow easy access to the quality control flags.
}
\section{Functions}{
\itemize{
\item \code{get_tech_failure}: get the technical failure flags for all samples.

\item \code{get_low_depth}: get the low depth failure flags for all samples.

\item \code{get_low_expression}: get the low expression failure flags for all genes.
}}

\examples{
object <- HermesData(summarized_experiment)

head(get_tech_failure(object))

head(get_low_depth(object))

head(get_low_expression(object))

}
