% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_snomed}
\alias{get_snomed}
\title{get SNOMED CT codes}
\usage{
get_snomed(version = "int")
}
\arguments{
\item{version}{a string, either "int" (international) or "uk"}
}
\value{
a data.table
}
\description{
get SNOMED CT codes
}
