% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{get_tech_failure}
\alias{get_tech_failure}
\title{Get Technical Failure Flags}
\usage{
get_tech_failure(object)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}
}
\value{
Named logical vector containing the technical failure flags for all samples.
}
\description{
Getter function which allows easy access to the technical failure flags.
}
\examples{
object <- HermesData(summarized_experiment)
get_tech_failure(object)

}
