% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrdSession.R
\name{getWrd}
\alias{getWrd}
\title{Get Current Word Session}
\usage{
getWrd(create = TRUE)
}
\arguments{
\item{create}{Logical; if \code{TRUE}, a new Word instance is created
when none is available.}
}
\value{
A Word COM object or \code{NULL}.
}
\description{
Retrieves the current Word COM object. If no valid session exists,
a new one can be created.
}
