% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlSession.R
\name{getXl}
\alias{getXl}
\title{Get Current Excel Session}
\usage{
getXl(create = TRUE)
}
\arguments{
\item{create}{Logical; if \code{TRUE}, a new Excel instance is
created when none is available.}
}
\value{
An Excel COM object or \code{NULL}.
}
\description{
Retrieves the current Excel COM object. If no valid session exists,
a new one can optionally be created.
}
