% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_all_duplicated}
\alias{h_all_duplicated}
\title{Finding All Duplicates in Vector}
\usage{
h_all_duplicated(x)
}
\arguments{
\item{x}{a (generalized, see \code{\link[base]{is.vector}}) vector, a data
    frame,  an array or \code{NULL}.}
}
\value{
Logical vector flagging all occurrences of duplicate values as \code{TRUE}.
}
\description{
The difference here to \code{\link[=duplicated]{duplicated()}} is that also the first occurrence
of a duplicate is flagged as \code{TRUE}.
}
\examples{
h_all_duplicated(c("a", "a", "b"))
duplicated(c("a", "a", "b"))
}
