% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{h_cpm}
\alias{h_cpm}
\title{Counts per Million (CPM) Normalization}
\usage{
h_cpm(object, control = control_normalize())
}
\arguments{
\item{object}{(\code{HermesData}) \cr input.}

\item{control}{(\code{list}) \cr list of settings used to perform the normalization procedure.}
}
\value{
A numeric matrix with normalized counts using the CPM method.
}
\description{
Counts per Million (CPM) Normalization
}
\examples{
h <- HermesData(summarized_experiment)
cont <- control_normalize()
counts_cpm <- h_cpm(h, cont)
str(counts_cpm)
}
