% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{h_diff_expr_deseq2}
\alias{h_diff_expr_deseq2}
\title{Helper function to use DESeq2 method}
\usage{
h_diff_expr_deseq2(object, design)
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{design}{(\code{matrix})\cr design matrix to run DESeq2 analysis.}
}
\value{
A data frame with DESeq2 analysis results.
}
\description{
This creates a data frame with folder change and p-values for all genes
}
