% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{h_diff_expr_limma}
\alias{h_diff_expr_limma}
\title{Helper function to use limma-voom method}
\usage{
h_diff_expr_limma(object, design)
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{design}{(\code{matrix})\cr design matrix to run limma-voom analysis.}
}
\value{
A data frame with limma-voom analysis results.
}
\description{
This creates a data frame with folder change and p-values for all genes
}
