% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{h_get_annotation_biomart}
\alias{h_get_annotation_biomart}
\title{Get Annotations from Biomart}
\usage{
h_get_annotation_biomart(gene_ids, id_var, mart)
}
\arguments{
\item{gene_ids}{(\code{character})\cr gene IDs.}

\item{id_var}{(\code{string})\cr gene ID variable, e.g. \code{entrezgene_id}.}

\item{mart}{(\code{Mart})\cr given \code{\link[biomaRt:Mart-class]{biomaRt::Mart}} object.}
}
\value{
A data frame with columns:
\itemize{
\item \code{id_var} (depending on what was used)
\item \code{hgnc_symbol}
\item \code{entrezgene_description}
\item \code{chromosome_name}
\item \code{start_position}
\item \code{end_position}
\item \code{refseq_mrna}
\item \code{refseq_peptide}
}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Helper function to query annotations from \code{biomaRt}, for cleaned up gene IDs of
a specific ID variable and given \code{\link[biomaRt:Mart-class]{biomaRt::Mart}}.
}
\examples{
\dontrun{
mart <- biomaRt::useMart("ensembl", dataset = "hsapiens_gene_ensembl")
h_get_annotation_biomart(c("11185", "10677"), id_var = "entrezgene_id", mart = mart)
}
}
