% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{h_low_depth_flag}
\alias{h_low_depth_flag}
\title{Quality Control: Low Depth Flag}
\usage{
h_low_depth_flag(object, control = control_quality())
}
\arguments{
\item{object}{(\code{HermesData}) \cr input.}

\item{control}{(\code{list}) \cr list of settings used to perform the quality control procedure.}
}
\value{
A logical vector indicating whether a sample in HermesData object has low average read depth.
}
\description{
Quality Control: Low Depth Flag
}
\examples{
object <- HermesData(summarized_experiment)
result <- h_low_depth_flag(object)
control <- control_quality(min_depth = 5)
result <- h_low_depth_flag(object, control)
head(result)

}
