% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{h_map_pos}
\alias{h_map_pos}
\title{Helper Function For Matching Map Values to Names}
\usage{
h_map_pos(names, map)
}
\arguments{
\item{names}{(\code{character})\cr original names.}

\item{map}{(named \code{character})\cr the mapping vector from old (value) to new
(name) names. All values must be included in \code{names}.}
}
\value{
Integer vector of the positions of the \code{map} values in the \code{names}.
}
\description{
This is used by the \code{\link{rename}} method. It wraps the assertions and the
matching used several times.
}
\examples{
h_map_pos(c("a", "b"), c(d = "b"))
}
