% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_cor_samplevar.R
\name{h_pca_var_rsquared}
\alias{h_pca_var_rsquared}
\title{Calculation of R2}
\usage{
h_pca_var_rsquared(pca, x)
}
\arguments{
\item{pca}{(\code{matrix})\cr A matrix comprising of principal components matrix generated by \code{\link[=calc_pca]{calc_pca()}} function.
function.}

\item{x}{(\code{vector})\cr A vector with one sample variable from \link{AnyHermesData} object.}
}
\value{
A vector with R2 values for each principal component for the sample variable of interest from \link{AnyHermesData}
object.
}
\description{
This function calculates R2 between one sample variable from \link{AnyHermesData} and Principal Component matrix.
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
 add_quality_flags() \%>\%
 filter() \%>\%
 normalize()
pca <- calc_pca(object)$x
x <- colData(object)$LowDepthFlag
r2 <- h_pca_var_rsquared(pca, x)
 
}
