% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{h_rpkm}
\alias{h_rpkm}
\title{Reads per Kilobase per Million (RPKM) Normalization}
\usage{
h_rpkm(object, control = control_normalize())
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input object.}

\item{control}{(\code{list})\cr list of settings used to perform the normalization procedure.}
}
\value{
A numeric matrix with normalized counts using the RPKM method.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This helper function calculates the RPKM normalized counts.
}
\note{
To be used with the \code{normalize()} function.
}
\examples{
h <- HermesData(summarized_experiment)
cont <- control_normalize(log = FALSE, lib_sizes = rep(1e6L, 20))
counts_rpkm <- h_rpkm(h, cont)
str(counts_rpkm)
}
