% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{h_tech_failure_flag}
\alias{h_tech_failure_flag}
\title{Quality Control: Technical Failure Flag}
\usage{
h_tech_failure_flag(object, threshold.corr = 0.5)
}
\arguments{
\item{object}{(\code{HermesData}) \cr input.}

\item{threshold.corr}{(\code{number})\cr threshold correlation parameter to perform the quality control procedure.}
}
\value{
A logical vector indicating whether each sample in HermesData object has failed the technical failure control.
}
\description{
Quality Control: Technical Failure Flag
}
\examples{
object <- HermesData(summarized_experiment)
result <- h_tech_failure_flag(object)
result <- h_tech_failure_flag(object, threshold.corr = 0.35)

}
