% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{h_tpm}
\alias{h_tpm}
\title{Transcripts per Million (TPM) Normalization}
\usage{
h_tpm(object, control)
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{control}{(\code{list})\cr list of settings used to perform the normalization procedure.}
}
\value{
Matrix of TPM normalized gene counts per sample.
}
\description{
Transcripts per Million (TPM) Normalization
}
\examples{
h <- HermesData(summarized_experiment)
cont <- control_normalize()
counts_tpm <- h_tpm(h, cont)
str(counts_tpm)

}
