% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSpec-class.R
\name{h_unique_labels}
\alias{h_unique_labels}
\title{Creation of Unique Labels}
\usage{
h_unique_labels(ids, nms = NULL)
}
\arguments{
\item{ids}{(\code{character} or \code{NULL})\cr unique IDs.}

\item{nms}{(\code{character} or \code{NULL})\cr not necessarily unique names if provided.}
}
\value{
Character vector where empty names are replaced by the IDs and
non-unique names are made unique by appending the IDs in parentheses.
}
\description{
This helper function generates a set of unique labels given
unique IDs and not necessarily unique names.
}
\examples{
h_unique_labels(c("1", "2", "3"), c("A", "B", "A"))
h_unique_labels(NULL)
h_unique_labels(c("1", "2", "3"))
}
