% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{h_voom}
\alias{h_voom}
\title{VOOM Normalization}
\usage{
h_voom(object, control = control_normalize())
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{control}{(\code{list})\cr list of settings used to perform the normalization procedure.}
}
\value{
A numeric matrix with normalized counts using the VOOM method.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function calculates the VOOM normalized counts.
}
\examples{
h <- HermesData(summarized_experiment)
cont <- control_normalize()
counts_voom <- h_voom(h, cont)
str(counts_voom)
}
