% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{extraColDataNames}
\alias{extraColDataNames}
\alias{extraRowDataNames}
\title{Helper Functions for Summary and Show Methods}
\usage{
extraColDataNames(x, ...)

extraRowDataNames(x, ...)
}
\arguments{
\item{...}{additional parameters for \code{colData()} and \code{rowData()}.}

\item{object}{(\code{AnyHermesData})\cr input.}
}
\value{
A character vector.
}
\description{
Helper functions, \code{extraColDataNames()} and \code{extraRowDataNames()}, used by \code{summary()}
and \code{show()}.
}
\examples{
object <- HermesData(summarized_experiment)
extraColDataNames(object)
extraRowDataNames(object)
}
