% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hermes_explicit_na}
\alias{hermes_explicit_na}
\title{Conversion of Character to Factor Variables in a \code{DataFrame}. For categorical
and logical variables, the new factors include a new level with missing data
label}
\usage{
hermes_explicit_na(data, na_level = "<Missing>")
}
\arguments{
\item{data}{(\code{DataFrame})\cr input \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}}.}

\item{na_level}{(\code{string})\cr missing level to be used.}
}
\value{
The modified data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function makes character to factor Variables in a \code{DataFrame}
with explicit missing data level.
}
\details{
This is using \link[forcats:fct_explicit_na]{forcats::fct_explicit_na} to add explicit missing
data level.
}
\examples{
dat <- colData(summarized_experiment)
any(sapply(dat, is.character))
any(sapply(dat, is.logical))
dat_hermes_convert <- hermes_explicit_na(dat)
any(sapply(dat_hermes_convert, function(x) is.character(x) || is.logical(x)))
}
