% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{class}
\name{HermesDataCor-class}
\alias{HermesDataCor-class}
\alias{.HermesDataCor}
\alias{HermesDataCor}
\title{HermesData Correlation}
\description{
The \link{HermesDataCor} class is an extension of a \link{matrix} with additional quality flags.
}
\note{
The \code{cor()} function will return an object of \link{HermesDataCor} class.
}
\examples{
object <- HermesData(summarized_experiment)
cor(x = object)
# which will return an object of HermesDataCor. 
         
}
