% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{class}
\name{HermesDataTopGenes-class}
\alias{HermesDataTopGenes-class}
\alias{.HermesDataTopGenes}
\title{Derivation of Top Genes}
\arguments{
\item{object}{(\code{AnyHermedData})\cr input.}

\item{assay_name}{(\code{string})\cr name of the assay to use for the sorting of genes.}

\item{summary_fun}{(\code{function})\cr summary statistics function to apply to the assay resulting in
a vector with one number per gene.}

\item{n_top}{(\code{count} or \code{NULL})\cr selection criteria based on number of entries.}

\item{min_threshold}{(\code{number} or \code{NULL} )\cr selection criteria based on a minimum
summary statistics threshold.}
}
\value{
\code{HermesDataTopGenes}\cr class object.
}
\description{
This creates a \code{data.frame} containing specified summary across columns and
the names of the rows. The data frame is sorted in descending order and the
top entries matching the selection criteria are selected.
Note that exactly one of the settings \code{n_top} and \code{min_threshold} must be provided.
}
\examples{
result <- HermesData(summarized_experiment)
object <- top_genes(object = result)
object <- top_genes(result, n_top = NULL, min_threshold = 50000)
object <- top_genes(result, summary_fun = rowMax)

}
