% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a HERMES study}
\usage{
init(cdm, target_cohort, comparator_cohort, outcome_cohort)
}
\arguments{
\item{cdm}{A cdm_reference object}

\item{target_cohort}{Name of the target cohort table}

\item{comparator_cohort}{Name of the comparator cohort table}

\item{outcome_cohort}{Name of the outcome cohort table}
}
\value{
A hermes_study object
}
\description{
Initialize a HERMES study
}
