% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{lapply,MultiAssayExperiment-method}
\alias{lapply,MultiAssayExperiment-method}
\title{\code{lapply} method for \code{MultiAssayExperiment}}
\usage{
\S4method{lapply}{MultiAssayExperiment}(X, FUN, safe = TRUE, ...)
}
\arguments{
\item{X}{(\code{MultiAssayExperiment})\cr input.}

\item{FUN}{A function to be applied to each \code{SummarizedExperiment} in \code{X}.}

\item{safe}{(\code{logical})\cr whether this method should skip experiment
where the function fails. Set to \code{TRUE} by default.}

\item{...}{additional arguments.}
}
\value{
\code{MultiAssayExperiment} object with specified function applied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Apply a function on all experiments in an MAE.
}
