% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_assay_experiment}
\alias{multi_assay_experiment}
\title{Example \code{MultiAssayExperiment} Data}
\format{
A \link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment} object with .
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
multi_assay_experiment
}
\description{
This example \link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment} can be used to .
}
\keyword{datasets}
