% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlSession.R
\name{newXl}
\alias{newXl}
\title{Create a New Excel Session}
\usage{
newXl(visible = TRUE)
}
\arguments{
\item{visible}{Logical; whether the Excel application should
be visible.}
}
\value{
An Excel COM object of class \code{"COMIDispatch"}.
}
\description{
Starts a new Microsoft Excel instance via RDCOMClient.
The created instance is registered as the current default
Excel session.
}
\examples{
\dontrun{
xl <- newXl()
}

}
