% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Phenotype}
\usage{
phenotype(
  x,
  id_col,
  code_col,
  include,
  exclude = NULL,
  name = "overall",
  include_multi = "any",
  exclude_multi = "any"
)
}
\arguments{
\item{x}{a data.frame like object}

\item{id_col}{a string, the name of the id column}

\item{code_col}{a single string, or vector of strings, the diagnosis code column(s)}

\item{include}{a string or list of strings, a valid phenotype id (see \code{get_phenotypes()})}

\item{exclude}{a string or list of strings, a valid phenotype id (see \code{get_phenotypes()})}

\item{name}{a string, a name for the phenotype}

\item{include_multi}{a string, how to deal with multiple include phenotypes, either 'all' (all required) or 'any' (all required)}

\item{exclude_multi}{a string, how to deal with multiple exclude phenotypes, either 'all' (all required) or 'any' (all required)}
}
\value{
a data.table
}
\description{
Phenotype
}
