% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
The result of the corresponding function call.
}
\description{
See \code{\link[magrittr]{\%>\%}} for details.
}
\examples{
hermes_data \%>\%
  filter() \%>\%
  normalize() \%>\%
  summary()
}
\keyword{internal}
