% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot,HermesDataCor,ANY-method}
\alias{plot,HermesDataCor,ANY-method}
\alias{plot}
\title{Plot Function}
\usage{
\S4method{plot}{HermesDataCor,ANY}(
  x,
  flag_colors = c(`FALSE` = "green", `TRUE` = "red"),
  cor_colors = circlize::colorRamp2(c(0, 0.5, 1), c("red", "yellow", "green")),
  ...
)
}
\arguments{
\item{x}{(\code{HermesDataCor})\cr object of correlation values to plot.}

\item{flag_colors}{(\code{character})\cr a vector that specifies the colors}

\item{cor_colors}{(\code{string})\cr the correlation coefficient (or covariance) to be computed, either "pearson", "kendall", or "spearman".}

\item{...}{arguments to be passed to}
}
\value{
A \link{HermesDataCor} object with calculated correlations and QC flags (technical failure and low depth).
}
\description{
Plot Function
}
\examples{
object <- HermesData(summarized_experiment)
result <- cor(x = object)
plot(result)
plot(result, show_column_names = FALSE, show_row_names = FALSE)

}
