% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_codes.R
\name{plot_code_overlap}
\alias{plot_code_overlap}
\title{Plot phenotype code overlap}
\usage{
plot_code_overlap(
  pheno_ids = c("PH_HF_HERMES_3.0", "PH25", "PH129"),
  types = c("ICD10 codes", "ICD9 codes", "OPCS4 codes", "Read codes v2", "Med codes",
    "SNOMED  CT codes", "OXMIS codes")
)
}
\arguments{
\item{pheno_ids}{a list of strings, valid phenotype IDs}

\item{types}{a vector of strings, one or more of c("ICD10 codes", "ICD9 codes", "OPCS4 codes", "Read codes v2", "Med codes", "SNOMED  CT codes", "OXMIS codes")}
}
\value{
a plot
}
\description{
Plot phenotype code overlap
}
