% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot,HermesDataCor,ANY-method}
\alias{plot,HermesDataCor,ANY-method}
\alias{plot_cor}
\title{Heatmap of Sample Correlations}
\usage{
\S4method{plot}{HermesDataCor,ANY}(
  x,
  y,
  flag_colors = c(`FALSE` = "green", `TRUE` = "red"),
  cor_colors = circlize::colorRamp2(c(0, 0.5, 1), c("red", "yellow", "green")),
  ...
)
}
\arguments{
\item{x}{(\code{HermesDataCor})\cr what to plot.}

\item{y}{not used.}

\item{flag_colors}{(named \code{character})\cr a vector that specifies the colors for \code{TRUE} and \code{FALSE}
flag values.}

\item{cor_colors}{(\code{function})\cr color scale function for the correlation values in the heatmap,
produced by \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}.}

\item{...}{other arguments to be passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}
}
\value{
The \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap} object with the heatmap.
}
\description{
This plot method uses the \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}} function
to plot the correlations between samples saved in a \link{HermesDataCor} object.
}
\examples{
object <- HermesData(summarized_experiment)
result <- calc_cor(object)
plot(result)
plot(result, show_column_names = FALSE, show_row_names = FALSE)

}
