% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot,HermesDataPca,ANY-method}
\alias{plot,HermesDataPca,ANY-method}
\alias{plot_pca}
\title{PCA Plot}
\usage{
\S4method{plot}{HermesDataPca,ANY}(x, y, ...)
}
\arguments{
\item{x}{(\code{HermesDataPca})\cr what to plot.}
}
\value{
The \code{ggplot} object with the PCA plot.
}
\description{
This plot method uses \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} function with the corresponding method
from the \code{ggfortify} package to plot the results of a principal components analysis
saved in a \link{HermesDataPca} object.
}
\examples{
object <- HermesData(summarized_experiment)
result <- calc_pca(object)
plot(result)
plot(result, variance_percentage = FALSE)
plot(result, label = TRUE)

}
