% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{prefix}
\alias{prefix}
\title{Prefix Accessor}
\usage{
prefix(object, ...)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{...}{additional arguments.}
}
\value{
The \code{prefix} slot contents.
}
\description{
Generic function to access the prefix from an object.
}
\examples{
a <- hermes_data
prefix(a)
}
