% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{rbind,HermesData-method}
\alias{rbind,HermesData-method}
\title{HermesData Methods}
\usage{
\S4method{rbind}{HermesData}(..., deparse.level = 1)
}
\arguments{
\item{args}{HermesData}
}
\value{
A HermesData
}
\description{
Combines HermesData objects with the same samples but differnet features of interest (rows in assays).
}
\examples{
hd <- rbind(hd1, hd2)
}
