% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{rbind2,HermesData,HermesData-method}
\alias{rbind2,HermesData,HermesData-method}
\alias{rbind}
\title{Row Binding of HermesData Objects}
\usage{
\S4method{rbind2}{HermesData,HermesData}(x, y, ...)
}
\arguments{
\item{x}{(\code{HermesData})\cr upper object.}

\item{y}{(\code{HermesData})\cr lower object.}
}
\value{
The combined \link{HermesData} object.
}
\description{
This method combines HermesData objects with the same samples but different
features of interest (rows in assays).
}
\details{
Note that this is used implicitly via the generic \code{\link[base:cbind]{base::rbind()}}
function, see \code{\link[methods:cbind2]{methods::rbind2()}} for more information about the
dispatching details.
}
\examples{
# todo

}
