% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrdSession.R
\name{setWrd}
\alias{setWrd}
\title{Set Current Word Session}
\usage{
setWrd(wrd)
}
\arguments{
\item{wrd}{A Word COM object (class \code{"COMIDispatch"}).}
}
\value{
Invisibly returns \code{wrd}.
}
\description{
Registers a Word COM object as the current default session.
}
