% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlSession.R
\name{setXl}
\alias{setXl}
\title{Set Current Excel Session}
\usage{
setXl(xl)
}
\arguments{
\item{xl}{An Excel COM object.}
}
\value{
Invisibly returns \code{xl}.
}
\description{
Registers an Excel COM object as the current default session.
}
