% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_economics}
\alias{simulate_economics}
\title{Simulate Economics}
\usage{
simulate_economics(
  traj_obj,
  time_horizon = 10,
  discount_rate = 0.03,
  n_samples = 100
)
}
\arguments{
\item{traj_obj}{A hermes_trajectories object}

\item{time_horizon}{Time horizon in years. Default is 10.}

\item{discount_rate}{Discount rate. Default is 0.03.}

\item{n_samples}{Number of PSA samples. Default is 100.}
}
\description{
Simulate Economics
}
