% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\docType{class}
\name{HermesDataSummary-class}
\alias{HermesDataSummary-class}
\alias{.HermesDataSummary}
\alias{HermesDataSummary}
\alias{summary,AnyHermesData-method}
\alias{show,HermesDataSummary-method}
\title{Summary Function for \code{AnyHermesData}}
\usage{
\S4method{summary}{AnyHermesData}(object)

\S4method{show}{HermesDataSummary}(object)
}
\arguments{
\item{object}{(\code{HermesDataSummary}) \cr input}
}
\description{
Summary Function for \code{AnyHermesData}
}
\section{Methods (by class)}{
\itemize{
\item \code{AnyHermesData}: A summary method for \link{AnyHermesData} object that creates a \link{HermesDataSummary} object.

\item \code{HermesDataSummary}: A show method prints summary description of \link{HermesDataSummary} object generated by
\code{\link[=summary]{summary()}} function.
}}

\examples{
object <- HermesData(summarized_experiment)
summary(object)
str(summary(object))

object <- HermesData(summarized_experiment)
result <- summary(object)
show(result)
result

}
