% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{class}
\name{HermesDataTopGenes-class}
\alias{HermesDataTopGenes-class}
\alias{.HermesDataTopGenes}
\alias{HermesDataTopGenes}
\alias{top_genes}
\title{Derivation of Top Genes}
\usage{
top_genes(
  object,
  assay_name = "counts",
  summary_fun = rowMeans,
  n_top = 10L,
  min_threshold = NULL
)
}
\arguments{
\item{object}{(\code{AnyHermedData})\cr input.}

\item{assay_name}{(\code{string})\cr name of the assay to use for the sorting of genes.}

\item{summary_fun}{(\code{function})\cr summary statistics function to apply to the assay resulting in
a numeric vector with one value per gene.}

\item{n_top}{(\code{count} or \code{NULL})\cr selection criteria based on number of entries.}

\item{min_threshold}{(\code{number} or \code{NULL} )\cr selection criteria based on a minimum
summary statistics threshold.}
}
\value{
A \link{HermesDataTopGenes} object.
}
\description{
This creates a \link{HermesDataTopGenes} object, which extends \code{data.frame}. It
contains two columns: \code{expression} containing the statistic values calculated
by \code{summary_fun} across columns, and \code{name} with the names of the rows. The
data frame is sorted in descending order of \code{expression} and only the top
entries according to the selection criteria are included.
}
\details{
Note that exactly one of the arguments \code{n_top} and \code{min_threshold} must be
provided.
}
\examples{
object <- HermesData(summarized_experiment)
top_genes(object)
top_genes(result, n_top = NULL, min_threshold = 50000)
top_genes(result, summary_fun = rowMax)

}
