% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdConst.R
\name{wdConst}
\alias{wdConst}
\title{Word Constants for RDCOMClient}
\format{
An object of class \code{"list"} containing named integer
constants (e.g., \code{wdCollapseEnd}, \code{wdAlignParagraphLeft}, etc.).
}
\usage{
wdConst
}
\description{
A list of Microsoft Word constants used for automation via
\code{RDCOMClient}. These constants correspond to enumerations
defined in the Word object model and are used when calling
methods on Word COM objects.
}
\details{
The constants are intended for use with Word COM interfaces, for example:
\preformatted{
wrd[["Selection"]]$Collapse(Direction = wdConst$wdCollapseEnd)
}

This avoids the need to manually define or remember numeric values
for Word enumerations.
}
\examples{
\dontrun{
# Collapse selection to end of document
wrd[["Selection"]]$Collapse(Direction = wdConst$wdCollapseEnd)

# Apply paragraph alignment
wrd[["Selection"]]$ParagraphFormat()$Alignment <- wdConst$wdAlignParagraphLeft
}

}
\seealso{
\code{\link{toWrd}}
}
\keyword{datasets}
