% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrdSession.R
\name{withWrd}
\alias{withWrd}
\title{Temporarily Use a Word Session}
\usage{
withWrd(wrd, expr)
}
\arguments{
\item{wrd}{A Word COM object.}

\item{expr}{An expression to evaluate.}
}
\value{
The result of the evaluated expression.
}
\description{
Evaluates an expression using a specified Word session as the default.
}
\examples{
\dontrun{
withWrd(newWrd(), {
  toWrd("Hello")
})
}

}
