% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlSession.R
\name{withXl}
\alias{withXl}
\title{Temporarily Use an Excel Session}
\usage{
withXl(xl, expr)
}
\arguments{
\item{xl}{An Excel COM object.}

\item{expr}{An expression to evaluate.}
}
\value{
The result of the evaluated expression.
}
\description{
Evaluates an expression using a specified Excel session as the
default.
}
