% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_in_mae}
\alias{wrap_in_mae}
\title{Wrap in MAE}
\usage{
wrap_in_mae(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{(\code{SummarizedExperiment})\cr input to create the MAE object from.}

\item{name}{(\code{string})\cr experiment name to use in the MAE for \code{x}.}
}
\value{
The MAE object with the only experiment being \code{x} having the given
\code{name}.
}
\description{
This helper function wraps \code{SummarizedExperiment} objects into an
a \code{MultiAssayExperiment} (MAE) object.
}
\examples{
mae <- wrap_in_mae(summarized_experiment)
mae[["summarized_experiment"]]
}
