% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrdBookmark.R
\name{wrdAddBookmark}
\alias{wrdAddBookmark}
\title{Add a Word Bookmark}
\usage{
wrdAddBookmark(name, wrd = NULL)
}
\arguments{
\item{name}{Bookmark name.}

\item{wrd}{A Word COM object. If \code{NULL}, the current
active Word session is used.}
}
\value{
Invisibly returns the created bookmark COM object.
}
\description{
Adds a new bookmark at the current cursor position
in the active Word document.
}
\examples{
\dontrun{
wrdAddBookmark("results")
}

}
