% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrdBookmark.R
\name{wrdBookmark}
\alias{wrdBookmark}
\title{Get a Word Bookmark}
\usage{
wrdBookmark(name, wrd = NULL)
}
\arguments{
\item{name}{Bookmark name.}

\item{wrd}{A Word COM object. If \code{NULL}, the current
active Word session is used.}
}
\value{
A Word bookmark COM object of class \code{"COMIDispatch"},
or \code{NULL} if the bookmark does not exist.
}
\description{
Retrieves a bookmark object from the active Word document.
}
\examples{
\dontrun{
bm <- wrdBookmark("intro")
}

}
