% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrdBookmark.R
\name{wrdDeleteBookmark}
\alias{wrdDeleteBookmark}
\title{Delete a Word Bookmark}
\usage{
wrdDeleteBookmark(name, wrd = NULL)
}
\arguments{
\item{name}{Bookmark name.}

\item{wrd}{A Word COM object. If \code{NULL}, the current
active Word session is used.}
}
\value{
Logical. Returns \code{TRUE} if the bookmark was deleted,
otherwise \code{FALSE}.
}
\description{
Deletes a bookmark from the active Word document.
}
\examples{
\dontrun{
wrdDeleteBookmark("intro")
}

}
