% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\docType{data}
\name{annotation,AnyHermesData-method}
\alias{annotation,AnyHermesData-method}
\alias{annotation}
\alias{.row_data_annotation_cols}
\alias{annotation<-,AnyHermesData,DataFrame-method}
\title{Annotation Accessor and Setter}
\format{
An object of class \code{character} of length 8.
}
\usage{
\S4method{annotation}{AnyHermesData}(object, ...)

.row_data_annotation_cols

\S4method{annotation}{AnyHermesData,DataFrame}(object) <- value
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr object to access the counts from.}

\item{...}{not used.}

\item{value}{(\code{matrix})\cr what should the counts assay be replaced with.}
}
\value{
The \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}} with the gene annotations:
\itemize{
\item \code{HGNC}
\item \code{HGNCGeneName}
\item \code{Chromosome}
\item \code{StartBP}
\item \code{EndBP}
\item \code{WidthBP}
\item \code{CanonicalTranscript}
\item \code{ProteinTranscript}
}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}

These methods access and set the gene annotations stored in a \code{\link{AnyHermesData}} object.
}
\note{
\itemize{
\item The returned column names are available in the exported
character vector \code{.row_data_annotation_cols}.
}

\itemize{
\item When trying to replace the annotation with completely missing values for any genes,
a warning will be given and the corresponding gene IDs will be saved in the
attribute \code{annotation.missing.genes}.
}
}
\examples{
object <- HermesData(summarized_experiment)
head(annotation(object))
}
\keyword{datasets}
