% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{annotation,AnyHermesData-method}
\alias{annotation,AnyHermesData-method}
\alias{annotation}
\alias{annotation<-,AnyHermesData,DataFrame-method}
\title{Annotation Accessor and Setter}
\usage{
\S4method{annotation}{AnyHermesData}(object, ...)

\S4method{annotation}{AnyHermesData,DataFrame}(object) <- value
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr object to access the counts from.}

\item{...}{not used.}

\item{value}{(\code{matrix})\cr what should the counts assay be replaced with.}
}
\value{
The \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}} with the gene annotations.
}
\description{
These methods access and set the gene annotations stored in a \code{\link{AnyHermesData}} object.
}
\examples{
object <- HermesData(summarized_experiment)
head(annotation(object))
}
