% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cor.R
\docType{class}
\name{correlate,AnyHermesData-method}
\alias{correlate,AnyHermesData-method}
\alias{calc_cor}
\alias{HermesDataCor-class}
\alias{.HermesDataCor}
\alias{HermesDataCor}
\alias{autoplot,HermesDataCor-method}
\title{Correlation between Sample Counts of \code{AnyHermesData}}
\usage{
\S4method{correlate}{AnyHermesData}(object, assay_name = "counts", method = "pearson", ...)

\S4method{autoplot}{HermesDataCor}(
  object,
  flag_colors = c(`FALSE` = "green", `TRUE` = "red"),
  cor_colors = circlize::colorRamp2(c(0, 0.5, 1), c("red", "yellow", "green")),
  ...
)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr object to calculate the correlation.}

\item{assay_name}{(\code{string})\cr the assay name where the counts are located in.}

\item{method}{(\code{string})\cr the correlation method, see \code{\link[stats:cor]{stats::cor()}} for details.}

\item{...}{other arguments to be passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}

\item{flag_colors}{(named \code{character})\cr a vector that specifies the colors for \code{TRUE} and \code{FALSE}
flag values.}

\item{cor_colors}{(\code{function})\cr color scale function for the correlation values in the heatmap,
produced by \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}.}
}
\value{
A \code{\link{HermesDataCor}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This calculates the correlation matrix between the sample vectors of counts from
a specified assay, as a \code{\link{HermesDataCor}} object which is an extension of a \code{\link{matrix}} with
additional quality flags in the slot \code{flag_data}: This contains the
\code{TechnicalFailureFlag} and \code{LowDepthFlag} columns describing the original input samples.
}
\section{Functions}{
\itemize{
\item \code{autoplot,HermesDataCor-method}: This plot method uses the \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}} function
to plot the correlations between samples saved in a \code{\link{HermesDataCor}} object.
}}

\examples{
object <- HermesData(summarized_experiment)
correlate(object)
result <- correlate(object, method = "pearson")
autoplot(result)
autoplot(result, show_column_names = FALSE, show_row_names = FALSE)
}
