% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_pca}
\alias{calc_pca}
\alias{`HermesDataPca-class`}
\title{Principal Components Analysis Calculation Function}
\usage{
calc_pca(object, assay_name = "counts")
}
\arguments{
\item{object}{(\code{AnyHermesData}) \cr input.}

\item{assay_name}{(\verb{Character string}) \cr Indicating the name of the assay
of interest, with possible options: "counts", "cpm", "tpm", "rpkm", "voom".
Default assay is "counts".}
}
\value{
A \link{HermesDataPca} object which is an extension of the \link[stats:prcomp]{stats::prcomp} class
to enable use of plot method for plotting PCA.
}
\description{
Perform principal components analysis of the gene count vectors across all
samples.
}
\details{
\itemize{
\item PCA should be performed after filtering out low quality genes and samples and
normalization.
\item In addition, genes with constant counts across all samples are excluded from
the analysis internally.
\item Centering and scaling is applied internally.
}
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
  add_quality_flags() \%>\%
  filter() \%>\%
  normalize()
result <- calc_pca(object, assay_name = "tpm")
summary(result)

}
