% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_with_newline}
\alias{cat_with_newline}
\title{Cats (prints) the objects with a final new line added}
\usage{
cat_with_newline(..., file = "")
}
\arguments{
\item{...}{arguments to cat}

\item{file}{file to write to, otherwise \code{stdout}}
}
\value{
objects with a final new line added.
}
\description{
#' @description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Use this function for simple objects as, for example,
it removes the "\link{1}" for scalars unlike print().

For non-simple objects like lists, use print to display
them (which is a generic method)

This imitates \code{cli::cat_line}.
}
\examples{
\dontrun{
cat_with_newline("hello", "world")
}
}
