% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_with_newline}
\alias{cat_with_newline}
\title{Concatenate and Print with Newline}
\usage{
cat_with_newline(..., file = "")
}
\arguments{
\item{...}{arguments to cat}
}
\value{
objects with a final new line added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function concatenate the representations like \code\code{\link[base:cat]{base::cat()}}
and prints them with newline.

For non-simple objects like lists, use print to display
them (which is a generic method)
}
\examples{
cat_with_newline("hello", "world")

}
\seealso{
This imitates \code\link[cli:cat_line]{cli::cat_line}.
}
