% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_with_newline}
\alias{cat_with_newline}
\title{Concatenate and Print with Newline}
\usage{
cat_with_newline(...)
}
\arguments{
\item{...}{inputs to concatenate.}
}
\value{
None, only used for the side effect of producing the concatenated output in the R console.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function concatenates inputs like \code{\link[=cat]{cat()}}
and prints them with newline.
}
\examples{
cat_with_newline("hello", "world")
}
\seealso{
This is similar to \code{\link[cli:cat_line]{cli::cat_line()}}.
}
