% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_proportion}
\alias{check_proportion}
\alias{assert_proportion}
\alias{test_proportion}
\alias{expect_proportion}
\title{Check for proportion}
\usage{
check_proportion(x, null.ok = FALSE)

assert_proportion(
  x,
  null.ok = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)

test_proportion(x, null.ok = FALSE)

expect_proportion(x, null.ok = FALSE, info = NULL, label = vname(x))
}
\arguments{
\item{x}{an object to check.}

\item{null.ok}{(\code{flag})\cr whether \code{x} may also be \code{NULL}.}

\item{.var.name}{(\code{string})\cr name of the checked object to print in
assertions; defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{(\code{AssertCollection} or \code{NULL})\cr collection to store
assertion messages, see \code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}.}

\item{info}{(\code{string})\cr extra information to be included in the
message for the \code{testthat} reporter, see \code{\link[testthat:expect_that]{testthat::expect_that()}}.}

\item{label}{(\code{string})\cr name of the checked object to print in
messages. Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}
}
\value{
\code{TRUE} if successful, otherwise a string with the error message.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Check whether \code{x} is a (single) proportion.
}
\examples{
check_proportion(0.25)
}
\seealso{
\code{\link{assertions}} for more details.
}
