% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colMeanZscores}
\alias{colMeanZscores}
\title{Generate gene signature using mean Z-score}
\usage{
colMeanZscores(x)
}
\arguments{
\item{x}{(\code{matrix})\cr containing numeric data.}
}
\value{
(\code{vector})\cr containing the mean Z-score.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function returns the Z-score from an assay stored as a \code{matrix}.
}
\examples{
object <- HermesData(summarized_experiment) |>
  add_quality_flags() |>
  filter() |>
  normalize() |>
  assay("counts")

colMeanZscores(object)
}
