% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colMeanZscores}
\alias{colMeanZscores}
\title{Mean Z-score Gene Signature}
\usage{
colMeanZscores(x)
}
\arguments{
\item{x}{(\code{matrix})\cr containing numeric data with genes in rows and samples
in columns, no missing values are allowed.}
}
\value{
A numeric vector containing the mean Z-score values for each
column in \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function returns the Z-score from an assay stored as a \code{matrix}.
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
  add_quality_flags() \%>\%
  filter() \%>\%
  normalize() \%>\%
  assay("counts")

colMeanZscores(object)
}
