% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\docType{class}
\name{connect_biomart}
\alias{connect_biomart}
\alias{ConnectionBiomart-class}
\alias{.ConnectionBiomart}
\alias{ConnectionBiomart}
\title{Connection to BioMart}
\usage{
connect_biomart(prefix = c("ENSG", "GeneID"))
}
\arguments{
\item{prefix}{(\code{string})\cr gene ID prefix.}
}
\value{
\code{\link{ConnectionBiomart}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{connect_biomart()} creates a connection object of class \code{\link{ConnectionBiomart}} which contains
the \code{biomaRt} object of class \code{\link[biomaRt:Mart-class]{biomaRt::Mart}} and the prefix of the object
which is used downstream for the query.
}
\details{
This connects to the Ensembl data base of BioMart for human genes.
}
\examples{
if (interactive()) {
  connection <- connect_biomart("ENSG")
}
}
