% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{control_normalize}
\alias{control_normalize}
\title{Settings for the normalize function}
\usage{
control_normalize(log = TRUE, lib_size = NULL, prior_count = 1)
}
\arguments{
\item{log}{(\code{flag})\cr whether log2 values are returned, otherwise original scale is used.}

\item{lib_size}{(\code{numeric})\cr library size, the default is colSums(y).}

\item{prior_count}{(\code{count})\cr average count to be added to each observation to avoid
taking log of zero, used only when log=TRUE.}
}
\value{
List with the above settings used to perform the normalization procedure.
}
\description{
Settings for the normalize function
}
\note{
Note, to be used with the normalize function.
}
\examples{
control <- control_normalize()
control1 <- control_normalize(log = FALSE, lib_size = 1000000L, prior_count = .5)
control2 <- control_normalize(lib_size = c(10L, 5L))
              
}
