% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{control_normalize}
\alias{control_normalize}
\title{Settings for the normalize function}
\usage{
control_normalize(log = TRUE, lib_size = NULL, prior_count = 1)
}
\arguments{
\item{log}{logical, if TRUE then log2 values are returned}

\item{lib_size}{library size, defaults to colSums(y)}

\item{prior_count}{average count to be added to each observation to avoid
taking log of zero, used only when log=TRUE}
}
\value{
Values for the settings used to perform the normalization procedure.
}
\description{
Settings for the normalize function
}
\note{
Note, to be used with the normalize function.
}
\examples{
control <- control_normalize()
control1 <- control_normalize(log = FALSE, lib_size = 1000000L, prior_count = 0.5)
              
}
