% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{control_normalize}
\alias{control_normalize}
\title{Control Settings for Counts Normalization}
\usage{
control_normalize(log = TRUE, lib_sizes = NULL, prior_count = 1)
}
\arguments{
\item{log}{(\code{flag})\cr whether \code{log2} values are returned, otherwise original scale is used.}

\item{lib_sizes}{(\code{numeric})\cr library sizes, the default is the vector with the sum of the
counts for each of the samples.}

\item{prior_count}{(\code{count})\cr average count to be added to each observation to avoid
taking log of zero, used only when \code{log = TRUE}.}
}
\value{
List with the above settings used to perform the normalization procedure.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}

This control function allows for easy customization of the normalization settings.
}
\note{
To be used with the \code{normalize()} function.
}
\examples{
control_normalize()
control_normalize(log = FALSE, lib_sizes = rep(1e6L, 20))
}
