% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{control_quality}
\alias{control_quality}
\title{Control for Specified Quality Flags}
\usage{
control_quality(
  min_cpm = 1,
  min_readcount_prop = 0.25,
  min_corr = 0.5,
  min_depth = NULL
)
}
\arguments{
\item{min_cpm}{(\code{number})\cr minimum cpm for each gene within the sample to pass, default is 1.}

\item{min_readcount_prop}{(\code{proportion})\cr minimum percentage of samples with acceptable read count of certain gene, default is 0.25.}

\item{min_corr}{(\code{proportion})\cr minimum correlation of RNAseq results, default is 0.5.}

\item{min_depth}{(\code{number})\cr minimum library depth, default is \code{NULL}, which means lower quartile minus 1.5 times
of the interquartile range is used as the threshold.}
}
\value{
List with the above criteria to flag observations.
}
\description{
Control for Specified Quality Flags
}
\note{
To be used with the \code{add_quality_flags()} function.
}
\examples{
control_quality()
control_quality(min_cpm = 5, min_readcount_prop = .001, min_corr = .1, min_depth = 3)
              
}
