% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_cor_samplevar.R
\docType{class}
\name{HermesDataPcaCor-class}
\alias{HermesDataPcaCor-class}
\alias{.HermesDataPcaCor}
\alias{HermesDataPcaCor}
\alias{correlate}
\alias{correlate,HermesDataPca-method}
\alias{autoplot,HermesDataPcaCor-method}
\title{Correlate}
\usage{
\S4method{correlate}{HermesDataPca}(object_pca, data)

\S4method{autoplot}{HermesDataPcaCor}(
  object,
  cor_colors = circlize::colorRamp2(c(-0.5, -0.25, 0, 0.25, 0.5, 0.75, 1), c("blue",
    "green", "purple", "yellow", "orange", "red", "brown")),
  cluster_columns = FALSE,
  ...
)
}
\arguments{
\item{object_pca}{(\code{HermesDataPca})\cr input. It can be generated using \code{\link[=calc_pca]{calc_pca()}} function on \link{AnyHermesData}.}

\item{data}{object (\code{AnyHermedData})\cr input.}

\item{object}{(\code{HermesDataPcaCor})\cr Generated using \link{correlate} method.}

\item{cor_colors}{(\code{function})\cr color scale function for the correlation values in the heatmap,
produced by \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}.}

\item{cluster_columns}{(\code{function})\cr Indicate whether re-ordering of columns is desired.}

\item{...}{other arguments to be passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}
}
\value{
A \link{HermesDataPcaCor} object with R2 values for all sample variables.
}
\description{
Method to generate correlation between all sample variables in \link{AnyHermesData} object and the principal components of
the samples.
}
\section{Methods (by class)}{
\itemize{
\item \code{HermesDataPca}: Method to obtain \link{HermesDataPcaCor} object with R2 values for all sample variables in
\link{AnyHermesData} object, giving correlations between all sample variables in \link{AnyHermesData} object and the
principal components of the samples.

\item \code{HermesDataPcaCor}: This plot method uses the \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}} function to plot the correlations between
samples variables in \link{AnyHermesData} object and the principal components of the samples saved in a \link{HermesDataPcaCor}
object.
}}

\examples{
object <- HermesData(summarized_experiment) \%>\%
 add_quality_flags() \%>\%
 filter() \%>\%
 normalize()
object_pca <- calc_pca(object)
result <- correlate(object_pca, object)

autoplot(result)

}
