% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{correlate}
\alias{correlate}
\title{Generic Function for Correlation Calculations}
\usage{
correlate(object, ...)
}
\arguments{
\item{object}{input of which the class will be used to decide the method.}

\item{...}{additional arguments.}
}
\value{
Corresponding object that contains the correlation results.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

New generic function to calculate correlations for one or two objects.
}
\seealso{
\link{pca_cor_samplevar} and \link{calc_cor} which are the methods included for this generic function.
}
