% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{correlate}
\alias{correlate}
\title{Generic Function for Correlation Calculations}
\usage{
correlate(object, ...)
}
\arguments{
\item{object}{input of which the class will be used to decide the method.}

\item{...}{additional arguments.}
}
\value{
Corresponding object that contains the correlation results.
}
\description{
Generic Function for Correlation Calculations
}
\seealso{
\link{pca_cor_samplevar} and \link{calc_cor} which are the methods included for this generic function.
}
