% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{counts,HermesData-method}
\alias{counts,HermesData-method}
\alias{counts}
\alias{counts<-,HermesData,matrix-method}
\title{Counts Accessor and Setter}
\usage{
\S4method{counts}{HermesData}(object)

\S4method{counts}{HermesData,matrix}(object) <- value
}
\arguments{
\item{object}{(\code{HermesData})\cr object to access the counts from.}

\item{value}{(\code{matrix})\cr what should the counts assay be replaced with.}
}
\value{
The counts assay.
}
\description{
These methods access and set the counts assay in a \link{HermesData} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = HermesData,value = matrix}: 
}}

\examples{
a <- HermesData(summarized_experiment)
result <- counts(a)
class(result)
head(result)

counts(a) <- counts(a) + 100L
head(counts(a))

}
