% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\docType{class}
\name{diff_expression}
\alias{diff_expression}
\alias{HermesDataDiffExpr-class}
\alias{.HermesDataDiffExpr}
\alias{HermesDataDiffExpr}
\title{Differential Expression Analysis}
\usage{
diff_expression(object, group, method = c("voom", "deseq2"))
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input. Note that this function only uses the
original counts for analysis, so this does not need to be normalized.}

\item{group}{(\code{string})\cr name of factor variable with 2 levels in \code{colData(object)}.
These 2 levels will be compared in the differential expression analysis.}

\item{method}{(\code{string})\cr method for differential expression analysis, see details below.}
}
\value{
A \code{\link{HermesDataDiffExpr}} object which is a data frame with the following columns for each gene
in the \code{\link{HermesData}} object:
\itemize{
\item \code{log2_fc} (estimate of the log2 fold change between the 2 levels of the
provided factor)
\item \code{stat} (the test statistic, which one depends on the method used)
\item \code{p_val} (the raw p-value),
\item \code{adj_p_val} (the adjusted p-value) values from differential expression analysis for each feature / gene .
}
}
\description{
This function performs differential expression analysis
using a method of preference.
}
\details{
Possible method choices are:
\itemize{
\item \code{voom}: uses \code{\link[limma:voom]{limma::voom()}}, see \code{\link[=h_diff_expr_voom]{h_diff_expr_voom()}} for details.
\item \code{deseq2}: uses \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}, see \code{\link[=h_diff_expr_deseq2]{h_diff_expr_deseq2()}} for details.
}
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
  add_quality_flags() \%>\%
  filter()
colData(object)$SEX <- factor(colData(object)$SEX) # nolint
res1 <- diff_expression(object, group = "SEX", method = "voom")
head(res1)
res2 <- diff_expression(object, group = "SEX", method = "deseq2")
head(res2)
}
