% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{diff_expression}
\alias{diff_expression}
\title{Generic Function for Differential Expression Analysis}
\usage{
diff_expression(object, group, method)
}
\description{
Generic Function for Differential Expression Analysis
}
\seealso{
\link{differential} which is the method defined for this generic function. And \link{h_diff_expr_voom} and
\link{h_diff_expr_deseq2} which are the helper functions included for this generic functions.
}
