% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_libsize_densities}
\alias{draw_libsize_densities}
\title{Density Plot of (Log) Counts Distributions}
\usage{
draw_libsize_densities(object, log = TRUE)
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{log}{(\code{flag})\cr should the counts be log transformed (log2).}
}
\value{
The \code{ggplot} object with the density plot.
}
\description{
This creates a density plot of the (log) counts distributions of the \link{HermesData} object where each line
on the plot corresponds to a sample.
}
\examples{
result <- HermesData(summarized_experiment)
draw_libsize_densities(result)
draw_libsize_densities(result, FALSE)

}
