% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_libsize_hist}
\alias{draw_libsize_hist}
\title{Histogram of Library Sizes}
\usage{
draw_libsize_hist(object, bins)
}
\arguments{
\item{object}{A \link{HermesData} object.}

\item{bins}{Number of evenly distributed groups desired.}
}
\description{
Histogram of Library Sizes
}
\examples{
result <- hermes:::.HermesData(summarized_experiment)
draw_libsize_hist(result, 30)

}
