% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_libsize_qq}
\alias{draw_libsize_qq}
\title{Q-Q Plot of Library Sizes}
\usage{
draw_libsize_qq(object, color = "grey", linetype = "dashed")
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{color}{(\code{string})\cr color of Q-Q line.}

\item{linetype}{(\code{string})\cr linetype of  Q-Q line.}
}
\value{
The \code{ggplot} object with the Q-Q Plot.
}
\description{
This creates a Q-Q plot of the library sizes of the \link{HermesData} object.
}
\examples{
result <- HermesData(summarized_experiment)
draw_libsize_qq(result)
draw_libsize_qq(result, color = "blue", linetype = "solid")
# We can also add sample name as label.
draw_libsize_qq(result) + geom_text(label = colnames(result), stat = "qq")
draw_libsize_qq(result, color = "blue", linetype = "solid") + geom_text(label = colnames(result), stat = "qq")

}
