% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_set}
\alias{expression_set}
\title{Example \code{ExpressionSet} Data}
\format{
A \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} object with 20 samples covering 5085
features (Entrez gene IDs).
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
expression_set
}
\description{
This example data can be used to try out conversion of a \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}
object into a \link{HermesData} object. Note that this data does not yet have the
required columns in the \code{featureData} and \code{phenoData}.
}
\seealso{
summarized_experiment which contains the same data as
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}.
}
\keyword{datasets}
