% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_set}
\alias{expression_set}
\title{Example \code{ExpressionSet} Data}
\format{
A \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object with 20 samples covering 5085
features (Entrez gene IDs).
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
expression_set
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}

This example data can be used to try out conversion of a \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}}
object into a \code{\link{HermesData}} object.
}
\seealso{
summarized_experiment which contains similar data as
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}.
}
\keyword{datasets}
