% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_set}
\alias{expression_set}
\title{Example \code{ExpressionSet} Data}
\format{
A \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object with 20 samples covering 5085
features (\code{Entrez} gene IDs).
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
expression_set
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This example data can be used to try out conversion of a \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}}
object into a \code{\link{HermesData}} object.
}
\seealso{
\itemize{
\item \code{\link[SummarizedExperiment:makeSummarizedExperimentFromExpressionSet]{SummarizedExperiment::makeSummarizedExperimentFromExpressionSet()}} to convert into a
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}.
\item \code{\link{summarized_experiment}} which contains similar data already as a
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}.
}
}
\keyword{datasets}
