% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{extra_data_names}
\alias{extra_data_names}
\alias{extraColDataNames,AnyHermesData-method}
\alias{extraRowDataNames,AnyHermesData-method}
\title{Extra Variable Names Accessor Methods}
\usage{
\S4method{extraColDataNames}{AnyHermesData}(x, ...)

\S4method{extraRowDataNames}{AnyHermesData}(x, ...)
}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object.}

\item{...}{not used.}
}
\value{
The character vector with the additional variable names in either
\code{colData()} or \code{rowData()}.
}
\description{
The methods access the names of the variables in \code{colData()} and \code{rowData()} of
the object which are not required by design. So these can be additional sample or
patient characteristics, or gene characteristics.
}
\section{Functions}{
\itemize{
\item \code{extraColDataNames,AnyHermesData-method}: provides names of extra \code{colData()} variables.

\item \code{extraRowDataNames,AnyHermesData-method}: provides names of extra \code{rowData()} variables.
}}

\examples{
object <- HermesData(summarized_experiment)
extraColDataNames(object)
extraRowDataNames(object)
}
