% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{extra_data_names}
\alias{extra_data_names}
\alias{extraColDataNames}
\alias{extraColDataNames,AnyHermesData-method}
\alias{extraRowDataNames}
\alias{extraRowDataNames,AnyHermesData-method}
\title{Extra Variable Names Accessor Methods}
\usage{
extraColDataNames(x, ...)

\S4method{extraColDataNames}{AnyHermesData}(x, ...)

extraRowDataNames(x, ...)

\S4method{extraRowDataNames}{AnyHermesData}(x, ...)
}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object.}

\item{...}{not used.}
}
\value{
The character vector with the additional variable names in either
\code{colData()} or \code{rowData()}.
}
\description{
The methods access the names of the variables in \code{colData()} and \code{rowData()} of
the object which are not required by design. So these can be additional sample or
patient characteristics, or gene characteristics.
}
\examples{
object <- HermesData(summarized_experiment)
extraColDataNames(object)
extraRowDataNames(object)
}
