% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{filter}
\alias{filter}
\title{Filter \code{HeremsData} object using QC flags.}
\usage{
filter(object, ...)
}
\arguments{
\item{object}{input of which the class will be used to decide the method.}

\item{...}{additional arguments.}
}
\description{
Filter \code{HeremsData} object using QC flags.
}
