% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{filter,AnyHermesData-method}
\alias{filter,AnyHermesData-method}
\alias{filter}
\title{Filter \code{AnyHermesData} on Subset Passing Default QC Flags}
\usage{
\S4method{filter}{AnyHermesData}(x)
}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object to filter.}
}
\value{
The filtered \code{\link{AnyHermesData}} object.
}
\description{
This filters a \code{\link{AnyHermesData}} object using the default QC flags. That is,
only genes without low expression (\code{LowExpressionFlag}) and samples
without low depth (\code{LowDepthFlag}) or technical failure (\code{TechnicalFailureFlag})
remain in the returned filtered object.
}
\note{
The internal implementation cannot use the \code{\link[=subset]{subset()}} method since that
requires non-standard evaluation of arguments.
}
\examples{
a <- HermesData(summarized_experiment)
dim(a)
result <- filter(a)
dim(result)
}
