% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{filter}
\alias{filter}
\alias{filter,AnyHermesData-method}
\title{Filter \code{AnyHermesData} on Subset Passing Default QC Flags}
\usage{
filter(x, ...)

\S4method{filter}{AnyHermesData}(x, what = c("genes", "samples"))
}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object to filter.}

\item{what}{(\code{vector}) \cr specify \code{genes} and / or \code{samples} to apply filter}
}
\value{
The filtered \code{\link{AnyHermesData}} object.
}
\description{
This filters a \code{\link{AnyHermesData}} object using the default QC flags. That is,
only genes without low expression (\code{LowExpressionFlag}) and samples
without low depth (\code{LowDepthFlag}) or technical failure (\code{TechnicalFailureFlag})
remain in the returned filtered object.
}
\note{
The internal implementation cannot use the \code{\link[=subset]{subset()}} method since that
requires non-standard evaluation of arguments.
}
\examples{
a <- HermesData(summarized_experiment)
dim(a)
# Filter genes and samples on default QC flags
result <- filter(a)
dim(result)
# Filter only genes without low expression
result <- filter(a, what = "genes")
}
