% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{genes}
\alias{genes}
\alias{genes,AnyHermesData-method}
\title{Gene IDs Accessor}
\usage{
genes(object)

\S4method{genes}{AnyHermesData}(object)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}
}
\value{
The character vector with the gene IDs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Access the gene IDs, i.e. row names, of a \code{\link{AnyHermesData}} object with a
nicely named accessor method.
}
\examples{
a <- HermesData(summarized_experiment)
genes(a)
}
\seealso{
\code{\link[=samples]{samples()}} to access the sample IDs.
}
