% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{h_diff_expr_deseq2}
\alias{h_diff_expr_deseq2}
\title{DESeq2 Differential Expression Analysis}
\usage{
h_diff_expr_deseq2(object, design)
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{design}{(\code{matrix})\cr design matrix.}
}
\value{
A data frame with columns \code{log2_fc} (estimated log2 fold change),
\code{stat} (test statistic), \code{p_val} (raw p-value), \code{adj_p_pval} (adjusted p-value).
}
\description{
This helper functions performs the differential expression analysis with
\code{\link[DESeq2:DESeq]{DESeq2::DESeq()}} for a given \link{AnyHermesData} input and \code{design} matrix.
}
\examples{
object <- HermesData(summarized_experiment)
design <- model.matrix(~ SEX, colData(object))
result <- h_diff_expr_deseq2(object, design)
head(result)

}
