% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{h_diff_expr_voom}
\alias{h_diff_expr_voom}
\title{\code{limma}/\code{voom} Differential Expression Analysis}
\usage{
h_diff_expr_voom(object, design)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{design}{(\code{matrix})\cr design matrix.}
}
\value{
A data frame with columns \code{log2_fc} (estimated log2 fold change),
\code{stat} (moderated t-statistic), \code{p_val} (raw p-value), \code{adj_p_pval} (Benjamini-Hochberg
adjusted p-value).
}
\description{
This helper functions performs the differential expression analysis with the \code{voom}
method from the \code{limma} package (via \code{\link[limma:voom]{limma::voom()}}, \code{\link[limma:lmFit]{limma::lmFit()}} and \code{\link[limma:ebayes]{limma::eBayes()}})
for given counts in a \link{AnyHermesData} object and a corresponding \code{design} matrix.
}
\examples{
object <- HermesData(summarized_experiment)
design <- model.matrix(~SEX, colData(object))
result <- h_diff_expr_voom(object, design)
head(result)
}
\references{
\insertRef{limma_package}{hermes}

\insertRef{voom_method}{hermes}
}
