% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{h_low_expression_flag}
\alias{h_low_expression_flag}
\title{Quality Control: Low Expression Flag}
\usage{
h_low_expression_flag(object, control = control_quality())
}
\arguments{
\item{object}{(\code{HermesData}) \cr input.}

\item{control}{(\code{list}) \cr list of settings used to perform the quality control procedure.}
}
\value{
A logical vector indicating whether each gene in HermesData object has low expression.
}
\description{
Quality Control: Low Expression Flag
}
\examples{
object <- HermesData(summarized_experiment)
result <- h_low_expression_flag(object)
control <- control_quality(min_cpm = 5, min_readcount_prop = 0.5)
result <- h_low_expression_flag(object, control)
head(result)

}
