% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_parens}
\alias{h_parens}
\title{Parenthesize a String}
\usage{
h_parens(x)
}
\arguments{
\item{x}{(\code{string})\cr input which should be parenthesized.}
}
\value{
String with parentheses, except when \code{x} is a blank string
in which case it is returned unaltered.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function adds parentheses around a string.
}
\examples{
h_parens("bla")
h_parens("")
}
