% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_parens}
\alias{h_parens}
\title{Parenthesize a Character Vector}
\usage{
h_parens(x)
}
\arguments{
\item{x}{(\code{character})\cr inputs which should be parenthesized.}
}
\value{
Character vector with parentheses, except when \code{x} is a blank string
in which case it is returned unaltered.
}
\description{
This helper function adds parentheses around each element of a character
vector.
}
\examples{
h_parens("bla")
h_parens("")
h_parens(c("bla", "bli"))
}
