% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_cor_samplevar.R
\name{h_pca_df_r2_matrix}
\alias{h_pca_df_r2_matrix}
\title{Generate R2 values for all sample variables}
\usage{
h_pca_df_r2_matrix(pca, df)
}
\arguments{
\item{pca}{(\code{matrix})\cr A matrix comprising of principal components matrix generated by \code{\link[=calc_pca]{calc_pca()}} function.}

\item{df}{(\code{dataframe})\cr A dataframe with sample data \link{colData} from \link{AnyHermesData} object.}
}
\value{
A matrix with R2 values for all sample variables in \link{AnyHermesData} object.
}
\description{
This function processes sample variables from \link{AnyHermesData} and the corresponding principal component matrix, and
then generates R2 values for all sample variables in \link{AnyHermesData} object, giving correlations between all sample
variables in \link{AnyHermesData} object and the principal components of the samples.
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
 add_quality_flags() \%>\%
 filter() \%>\%
 normalize()
pca <- calc_pca(object)$x
r2_all <- h_pca_df_r2_matrix(pca, as.data.frame(colData(object)))

}
