% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_cor_samplevar.R
\name{h_pca_df_r2_matrix}
\alias{h_pca_df_r2_matrix}
\title{Calculation of R2 Matrix between Sample Variables and Principal Components}
\usage{
h_pca_df_r2_matrix(pca, df)
}
\arguments{
\item{pca}{(\code{matrix})\cr comprises principal components generated by \code{\link[=calc_pca]{calc_pca()}}.}

\item{df}{(\code{dataframe})\cr \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::colData()}} of \link{AnyHermesData} object.}
}
\value{
A matrix with R2 values for all combinations of sample variables and principle
components.
}
\description{
This function processes sample variables from \link{AnyHermesData} and the
corresponding principal components matrix, and then generates the matrix of R2 values.
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
  add_quality_flags() \%>\%
  filter() \%>\%
  normalize()
pca <- calc_pca(object)$x
r2_all <- h_pca_df_r2_matrix(pca, as.data.frame(colData(object)))
}
