% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_cor_samplevar.R
\name{h_pca_var_rsquared}
\alias{h_pca_var_rsquared}
\title{Calculation of R2 between Sample Variable and Principal Components}
\usage{
h_pca_var_rsquared(pca, x)
}
\arguments{
\item{pca}{(\code{matrix})\cr principal components matrix generated by \code{\link[=calc_pca]{calc_pca()}}.}

\item{x}{(\code{vector})\cr values of one sample variable from \link{AnyHermesData} object.}
}
\value{
A vector with R2 values for each principal component.
}
\description{
This helper function calculates R2 values between one sample variable from \link{AnyHermesData}
and all Principal Components (PCs) separately (one linear model is fit for each PC).
}
\examples{
object <- HermesData(summarized_experiment) \%>\%
  add_quality_flags() \%>\%
  filter() \%>\%
  normalize()
pca <- calc_pca(object)$x
x <- colData(object)$LowDepthFlag
r2 <- h_pca_var_rsquared(pca, x)
 
}
