% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{h_rpkm}
\alias{h_rpkm}
\title{Helper function to get Reads per Kilobase per Million (RPKM)}
\usage{
h_rpkm(object, control)
}
\arguments{
\item{object}{(\code{HermesData})\cr input object.}

\item{control}{(\code{list})\cr list of settings used to perform the normalization procedure.}
}
\value{
A numeric matrix with normalized counts using the RPKM method.
}
\description{
Helper function to get Reads per Kilobase per Million (RPKM)
}
\note{
To be used with the \code{normalize()} function.
}
\examples{
h <- HermesData(summarized_experiment)
cont <- control_normalize(log = FALSE, lib_sizes = rep(1e6L, 20))
counts_rpkm <- h_rpkm(h, cont)
str(counts_rpkm)

}
