% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{h_tech_failure_flag}
\alias{h_tech_failure_flag}
\title{Quality Control: Technical Failure Flag}
\usage{
h_tech_failure_flag(object, control = control_quality())
}
\arguments{
\item{object}{(\code{HermesData}) \cr input.}

\item{control}{(\code{list}) \cr list of settings used to perform the quality control procedure.}
}
\value{
A logical vector indicating whether each sample in HermesData object has failed the technical failure control.
}
\description{
Quality Control: Technical Failure Flag
}
\examples{
object <- HermesData(summarized_experiment)
result <- h_tech_failure_flag(object)
control <- control_quality(min_corr = 0.35)
result <- h_tech_failure_flag(object, control)

}
