% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{h_tpm}
\alias{h_tpm}
\title{Transcripts per Million (TPM) Normalization}
\usage{
h_tpm(object, control = control_normalize())
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{control}{(\code{list})\cr list of settings used to perform the normalization procedure.}
}
\value{
A numeric matrix with normalized counts using the TPM method.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This helper function calculates the TPM normalized counts.
}
\examples{
h <- HermesData(summarized_experiment)
cont <- control_normalize()
counts_tpm <- h_tpm(h, cont)
str(counts_tpm)
}
