% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{h_tpm}
\alias{h_tpm}
\title{Transcripts per Million (TPM) Normalization}
\usage{
h_tpm(object, control = control_normalize())
}
\arguments{
\item{object}{(\code{HermesData})\cr input.}

\item{control}{(\code{list})\cr list of settings used to perform the normalization procedure.}
}
\value{
A numeric matrix with normalized counts using the TPM method.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}

This helper function calculates the TPM normalized counts.
}
\examples{
h <- HermesData(summarized_experiment)
cont <- control_normalize()
counts_tpm <- h_tpm(h, cont)
str(counts_tpm)
}
