% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{hermes-package}
\alias{hermes}
\alias{hermes-package}
\title{\code{hermes} Package}
\description{
\code{hermes} facilitates preprocessing, analyzing, and reporting of RNA-seq data.
}
\author{
\strong{Maintainer}: NEST \email{nestcicd@roche.com}

}
