% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{hermes-package}
\alias{hermes}
\alias{hermes-package}
\title{\code{hermes} Package}
\description{
\code{hermes} facilitates preprocessing, analyzing, and reporting of RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.roche.com/NEST/hermes/}
  \item Report bugs at \url{https://github.roche.com/NEST/hermes/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{nestcicd@roche.com}

}
