% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{hermes-package}
\alias{hermes}
\alias{hermes-package}
\title{\code{hermes} Package}
\description{
\code{hermes} facilitates preprocessing, analyzing, and reporting of RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/hermes/}
  \item Report bugs at \url{https://github.com/insightsengineering/hermes/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanés Bové \email{daniel.sabanes_bove@roche.com}

Authors:
\itemize{
  \item Namrata Bhatia
  \item Stefanie Bienert
  \item Benoit Falquet
  \item Haocheng Li
  \item Jeff Luong
  \item Lyndsee Midori Zhang \email{zhang.lyndsee@gene.com}
  \item Alex Richardson
  \item Simona Rossomanno
  \item Tim Treis
  \item Mark Yan
  \item Naomi Chang
  \item Chendi Liao \email{chendi.liao@roche.com}
  \item Carolyn Zhang
  \item Joseph N. Paulson \email{paulson.joseph@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
