% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermes-package.R
\docType{package}
\name{hermes-package}
\alias{hermes}
\alias{hermes-package}
\title{hermes: Health Economics and Outcomes Research Pipeline on OMOP CDM (Metapackage)}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

An R metapackage providing an end-to-end analytical ecosystem for Real-World Evidence (RWE) and Health Economics and Outcomes Research (HEOR). Unifies CohortUtilisation (HCRU), CohortCosts (direct medical expenditures), and CohortEconomics (propensity scores, Markov simulation, and Cost-Effectiveness Analysis) on OMOP CDM data. Developed by IOMED.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iomedhealth.github.io/hermes/}
  \item \url{https://www.iomed.health/}
  \item \url{https://github.com/iomedhealth/hermes}
}

}
\author{
\strong{Maintainer}: Gabriel Maeztu \email{gabriel@iomed.health}

Authors:
\itemize{
  \item Gabriel Maeztu \email{gabriel@iomed.health}
}

}
\keyword{internal}
