% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERMES-package.R
\docType{package}
\name{HERMES-package}
\alias{HERMES}
\alias{HERMES-package}
\title{HERMES: Health Economics and Outcomes Research Pipeline on OMOP CDM}
\description{
An R package ecosystem for Real-World Evidence (RWE) and Health Economics and Outcomes Research (HEOR) focused on Healthcare Resource Utilization (HCRU) and Cost-Effectiveness Analysis (CEA) on OMOP CDM data.
}
\keyword{internal}
