% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERMES-package.R
\docType{package}
\name{HERMES-package}
\alias{HERMES}
\alias{HERMES-package}
\title{HERMES: Health Economics and Outcomes Research Pipeline on OMOP CDM}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

An R package ecosystem for Real-World Evidence (RWE) and Health Economics and Outcomes Research (HEOR) focused on Healthcare Resource Utilization (HCRU) and Cost-Effectiveness Analysis (CEA) on OMOP CDM data. Developed by IOMED.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.iomed.health/}
  \item \url{https://github.com/iomedhealth/hermes}
}

}
\author{
\strong{Maintainer}: Gabriel Maeztu \email{gabriel@iomed.health}

Authors:
\itemize{
  \item Gabriel Maeztu \email{gabriel@iomed.health}
}

}
\keyword{internal}
