% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heRmes-package.R
\docType{package}
\name{heRmes-package}
\alias{heRmes}
\alias{heRmes-package}
\title{heRmes: Tools Used By The HERMES Consortium}
\description{
A collection of tools for phenotyping and processing heart failure GWAS data.
}
\author{
\strong{Maintainer}: Nicholas Sunderland \email{nicholas.sunderland@bristol.ac.uk} (\href{https://orcid.org/0000-0003-3625-630X}{ORCID})

}
\keyword{internal}
