% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hermes_data}
\alias{hermes_data}
\title{Example \code{HermesData} Data}
\format{
A \link{HermesData} object with 20 samples covering
5085 features (\code{Entrez} gene IDs).
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
hermes_data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This example \code{\link{HermesData}} is created from the underlying \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}
object by renaming descriptors to align with standard specification. It already
contains the required columns in \code{rowData} and \code{colData}.
}
\seealso{
\code{\link{summarized_experiment}} for the underlying \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}
object.
}
\keyword{datasets}
