% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-class.R
\docType{class}
\name{HermesData-class}
\alias{HermesData-class}
\alias{.HermesData}
\alias{HermesData}
\alias{RangedHermesData}
\alias{AnyHermesData}
\alias{RangedHermesData-class}
\alias{.RangedHermesData}
\alias{AnyHermesData-class}
\alias{HermesDataFromMatrix}
\title{\code{HermesData} and \code{RangedHermesData}}
\usage{
HermesData(object)

HermesDataFromMatrix(counts, ...)
}
\arguments{
\item{object}{(\code{SummarizedExperiment})\cr input to create \code{\link{HermesData}} from.
If this is a \code{RangedSummarizedExperiment}, then the result will be
\code{\link{RangedHermesData}}.}

\item{counts}{(\code{matrix})\cr counts to create the \code{\link{HermesData}} object from.}

\item{...}{additional arguments, e.g. \code{rowData}, \code{colData}, etc. passed to
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} internally. Note that if \code{rowRanges}
is passed instead of \code{rowData}, then the result will be a \code{\link{RangedHermesData}} object.}
}
\description{
The \code{\link{HermesData}} class is an extension of \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}
with additional validation criteria.
}
\details{
The additional criteria are:
\itemize{
\item The first assay must be \code{counts} containing non-missing, integer, non-negative values.
\item The following columns must be in \code{rowData}:
\itemize{
\item \code{HGNC}
\item \code{HGNCGeneName}
\item \code{GeneID}
\item \code{Chromosome}
\item \code{StartBP}
\item \code{EndBP}
\item \code{WidthBP}
\item \code{CanonicalTranscript}
\item \code{ProteinTranscript}
\item \code{LowExpressionFlag}
}
\item The following columns must be in \code{colData}:
\itemize{
\item \code{SampleID}
\item \code{LowDepthFlag}
\item \code{TechnicalFailureFlag}
}
\item The object must have row and column names.
}

Analogously, \code{\link{RangedHermesData}} is an extension of
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}} and has the same
additional validation requirements. Methods can be defined for both classes at the
same time with the \code{\link{AnyHermesData}} signature.

A \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object can be imported by using the
\code{\link[SummarizedExperiment:makeSummarizedExperimentFromExpressionSet]{SummarizedExperiment::makeSummarizedExperimentFromExpressionSet()}} function to
first convert it to a \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object before
converting it again into a \code{\link{HermesData}} object.
}
\section{Slots}{

\describe{
\item{\code{prefix}}{common prefix of the gene IDs (row names).}
}}

\note{
Note that we use \code{\link[S4Vectors:S4Vectors-internals]{S4Vectors::setValidity2()}} to define the validity
method, which allows us to turn off the validity checks in internal
functions where intermediate objects may not be valid within the scope of
the function.
}
\examples{
# Convert to `SummarizedExperiment` using the default naive range mapper.
se <- makeSummarizedExperimentFromExpressionSet(expression_set)
# Then convert to `HermesData`.
# Create objects starting from a `SummarizedExperiment.`
hermes_data <- HermesData(summarized_experiment)
hermes_data
ranged_summarized_experiment <- as(summarized_experiment, "RangedSummarizedExperiment")
ranged_hermes_data <- HermesData(ranged_summarized_experiment)
ranged_hermes_data
# Create objects from a matrix and additional arguments.
counts_matrix <- assay(summarized_experiment)
HermesDataFromMatrix(
  counts = counts_matrix,
  rowData = rowData(summarized_experiment),
  colData = colData(summarized_experiment)
)
}
