% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{isEmpty,SummarizedExperiment-method}
\alias{isEmpty,SummarizedExperiment-method}
\alias{isEmpty}
\title{Checking for Empty \code{SummarizedExperiment}}
\usage{
\S4method{isEmpty}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{(\code{SummarizedExperiment})\cr object to check.}
}
\value{
Flag whether the \code{object} is empty.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This method checks whether a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object is empty.
}
\examples{
isEmpty(summarized_experiment)
isEmpty(summarized_experiment[NULL, ])
isEmpty(hermes_data)
}
