% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{lapply,MultiAssayExperiment-method}
\alias{lapply,MultiAssayExperiment-method}
\alias{lapply}
\title{\code{lapply} method for \code{MultiAssayExperiment}}
\usage{
\S4method{lapply}{MultiAssayExperiment}(X, FUN, safe = TRUE, ...)
}
\arguments{
\item{X}{(\code{MultiAssayExperiment})\cr input.}

\item{FUN}{(\code{function}) to be applied to each experiment in \code{X}.}

\item{safe}{(\code{flag})\cr whether this method should skip experiments
where the function fails.}

\item{...}{additional arguments passed to \code{FUN}.}
}
\value{
\code{MultiAssayExperiment} object with specified function applied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Apply a function on all experiments in an MAE.
}
