% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{metadata}
\alias{metadata}
\title{Metadata Accessor and Setter}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object to access the metadata from.}

\item{value}{(\code{list})\cr the list to replace the current metadata with.}
}
\value{
The metadata which is a list.
}
\description{
These methods access or set the metadata in a \code{\link{AnyHermesData}} object.
}
\note{
Note that this just inherits \code{\link[S4Vectors:Annotated-class]{S4Vectors::metadata,Annotated-method()}}.
}
\examples{
a <- hermes_data
metadata(a)
metadata(a) <- list(new = "my metadata")
metadata(a)
}
