% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{metadata}
\alias{metadata}
\title{Metadata Accessor and Setter}
\arguments{
\item{x}{(\code{HermesData})\cr object to access the metadata from.}
}
\value{
The metadata which is a list.
}
\description{
These methods access or set the metadata in a \link{HermesData} object.
}
\note{
Note that this just inherits \code{\link[S4Vectors:Annotated-class]{S4Vectors::metadata,Annotated-method()}}.
}
\examples{
a <- HermesData(summarized_experiment)
metadata(a)
metadata(a) <- list(new = "my metadata")
metadata(a)

}
