% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_assay_experiment}
\alias{multi_assay_experiment}
\title{Example \code{MultiAssayExperiment} Data}
\format{
A \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}} object with 3 separate SE objects. The first SE object
contains 5 samples and covers 1000 features (Entrez gene IDs). The second SE object contains 9 samples with
2500 features. The third SE object contains 6 samples with 1300 features.
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
multi_assay_experiment
}
\description{
This example \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}} can be used as test data.
}
\keyword{datasets}
