% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_assay_experiment}
\alias{multi_assay_experiment}
\title{Example \code{MultiAssayExperiment} Data}
\format{
A \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}} object with 3 separate SE objects.
\itemize{
\item The first SE object contains 5 samples and covers 1000 features (Entrez gene IDs).
\item The second SE object contains 9 samples with 2500 features.
\item The third SE object contains 6 samples with 1300 features.
}
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
multi_assay_experiment
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

This example \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}} can be used as test data.
}
\keyword{datasets}
