% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_assay_experiment}
\alias{multi_assay_experiment}
\title{Example \code{MultiAssayExperiment} Data}
\format{
A \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}} object with 3 separate \code{\link{HermesData}}
objects.
\itemize{
\item The first object contains 5 samples and covers 1000 features (\code{Entrez} gene IDs).
\item The second object contains 9 samples with 2500 features.
\item The third object contains 6 samples with 1300 features.
}
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
multi_assay_experiment
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This example \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}} can be used as test data.
}
\keyword{datasets}
