% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopHeor-package.R
\docType{package}
\name{omopHeor-package}
\alias{omopHeor}
\alias{omopHeor-package}
\title{omopHeor: Health Economics and Outcomes Research Pipeline on OMOP CDM}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

An R metapackage providing an end-to-end analytical ecosystem for Real-World Evidence (RWE) and Health Economics and Outcomes Research (HEOR). Unifies 'CohortUtilisation' (Healthcare Resource Utilization), 'CohortCosts' (direct medical expenditures), and 'CohortEconomics' (propensity scores, Markov simulation, and Cost-Effectiveness Analysis) on Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM) data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iomedhealth.github.io/omopHeor/}
  \item \url{https://www.iomed.health/}
  \item \url{https://github.com/iomedhealth/omopHeor}
  \item Report bugs at \url{https://github.com/iomedhealth/omopHeor/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Maeztu \email{gabriel@iomed.health}

Authors:
\itemize{
  \item Gabriel Maeztu \email{gabriel@iomed.health}
}

}
\keyword{internal}
