% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Phenotype}
\usage{
phenotype(ids, codes, name, include, exclude = NULL)
}
\arguments{
\item{ids}{a vector of characters IDs}

\item{codes}{a vector of diagnosis codes}

\item{name}{a string, a name for the phenotype}

\item{include}{a string or list of strings, a valid phenotype id (see \code{get_phenotypes()})}

\item{exclude}{a string or list of strings, a valid phenotype id (see \code{get_phenotypes()})}
}
\value{
a data.table
}
\description{
Phenotype
}
