% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Phenotype}
\usage{
phenotype(
  x,
  id_col,
  include,
  exclude = NULL,
  code_cols = list(`ICD9 codes` = NULL, `ICD10 codes` = NULL, `OPCS4 codes` = NULL,
    `Read codes v2` = NULL, `Read codes v3` = NULL),
  gsub = NULL,
  name = "overall",
  include_multi = "any",
  exclude_multi = "any",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{a data.frame like object or file path readable by data.table::fread}

\item{id_col}{a string, the name of the id column}

\item{include}{a string or list of strings, a valid phenotype id (see \code{get_phenotypes()})}

\item{exclude}{a string or list of strings, a valid phenotype id (see \code{get_phenotypes()})}

\item{code_cols}{a list of named strings, the diagnosis and procedure code column(s). List elements must be one or more of
\code{icd9}, \code{icd10}, \code{opcs}, \code{readv2}, \code{readv3}}

\item{gsub}{a list of characters of length 3, pre-processing of phenotype codes and/or codes in \code{x}. E.g. c(".", "", "both"). NULL to turn off.
Element 1: a character, regex pattern.
Element 2: a character, replacement string.
Element 3: a character vector, one or more of: 'x' (apply to codes in \code{x} only), 'pheno' (apply to codes in pheno files only), 'both' (apply to both \code{x} and all pheno files), or a phenotype_id in \code{include} or \code{exclude} (apply to that phenotype).}

\item{name}{a string, a name for the phenotype}

\item{include_multi}{a string, how to deal with multiple include phenotypes, either 'all' (all required) or 'any' (all required)}

\item{exclude_multi}{a string, how to deal with multiple exclude phenotypes, either 'all' (all required) or 'any' (all required)}

\item{verbose}{a logical, print progress}

\item{...}{other parameters passed to gsub}
}
\value{
a data.table
}
\description{
Phenotype
}
