% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot,HermesDataPca,ANY-method}
\alias{plot,HermesDataPca,ANY-method}
\alias{plot_pca}
\title{PCA Plot}
\usage{
\S4method{plot}{HermesDataPca,ANY}(x, y, x_comp = 1, y_comp = 2, ...)
}
\arguments{
\item{x}{(\code{HermesDataPca})\cr what to plot.}

\item{y}{not used.}

\item{x_comp}{(\code{count})\cr principal component number used in x axis.}

\item{y_comp}{(\code{count})\cr principal component number used in y axis.}

\item{...}{additional arguments passed internally to \link[ggfortify:autoplot.pca_common]{ggfortify::autoplot.prcomp}.}
}
\value{
The \code{ggplot} object with the PCA plot.
}
\description{
This plot method uses \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} function with the corresponding method
from the \code{ggfortify} package to plot the results of a principal components analysis
saved in a \link{HermesDataPca} object.
}
\examples{
object <- HermesData(summarized_experiment)
result <- calc_pca(object)
plot(result)
plot(result, x_comp = 2, y_comp = 3)
plot(result, variance_percentage = FALSE)
plot(result, label = TRUE)

}
