% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{prefix}
\alias{prefix}
\title{Prefix Accessor}
\usage{
prefix(object, ...)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{...}{additional arguments.}
}
\value{
The \code{prefix} slot contents.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Generic function to access the prefix from an object.
}
\examples{
a <- HermesData(summarized_experiment)
prefix(a)
}
