% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{prefix}
\alias{prefix}
\title{Prefix Accessor}
\usage{
prefix(object, ...)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}

\item{...}{additional arguments.}
}
\value{
The \code{prefix} slot contents.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generic function to access the prefix from an object.
}
\examples{
a <- HermesData(summarized_experiment)
prefix(a)
}
