% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{query}
\alias{query}
\alias{query,character,Igis-method}
\title{Query Gene Annotations from a Connection}
\usage{
query(genes, connection)

\S4method{query}{character,Igis}(genes, connection)
}
\arguments{
\item{genes}{(\code{character})\cr gene IDs.}

\item{connection}{(\code{Igis} or other connection class)\cr data base connection object.}
}
\value{
A \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}} with the gene annotations.
}
\description{
This generic function is the interface for querying gene annotations from
a data base connection. Currently only the IGIS version is implemented.
}
\examples{
\dontrun{
object <- HermesData(summarized_experiment)
connection <- connect_igis(prefix(object))
result <- query(genes(object), connection)
head(result)
}
}
