% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{rbind}
\alias{rbind}
\title{Row Binding of HermesData Objects}
\arguments{
\item{...}{(\code{HermesData})\cr objects to row bind.}
}
\value{
The combined \link{HermesData} object.
}
\description{
This method combines HermesData objects with the same samples but different
features of interest (rows in assays).
}
\note{
Note that this just inherits
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::rbind,SummarizedExperiment-method()}}. When binding a
\link{HermesData} object with a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}
object, then the result will be a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} object (the more general
class).
}
\examples{
a <- b <- HermesData(summarized_experiment)
result <- rbind(a, b)
class(result)

result2 <- rbind(summarized_experiment, b)
class(result2)

}
