% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{addInpatients}
\alias{addHospitalizations}
\alias{addInpatient}
\alias{addEmergencyCare}
\alias{addEmergency}
\alias{addEmergencyVisits}
\alias{addOutpatientVisits}
\alias{addVisits}
\alias{addPrescriptions}
\alias{addProcedures}
\alias{computeHospitalizationCohorts}
\alias{compute_hospitalization_cohorts}
\alias{computeInfusionCohorts}
\alias{summariseUtilization}
\alias{tableUtilization}
\alias{plotUtilization}
\alias{addCosts}
\alias{summariseCosts}
\alias{tableCosts}
\alias{plotCosts}
\alias{init}
\alias{summarise_baseline}
\alias{extract_hcru}
\alias{extractHcru}
\alias{fit_ps}
\alias{adjust_ps}
\alias{assess_balance}
\alias{compile_trajectories}
\alias{simulate_economics}
\alias{run_cea}
\alias{plot_ceac}
\alias{plot_plane}
\alias{table_summary}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CohortCosts}{\code{\link[CohortCosts:addCosts]{addCosts()}}, \code{\link[CohortCosts:plotCosts]{plotCosts()}}, \code{\link[CohortCosts:summariseCosts]{summariseCosts()}}, \code{\link[CohortCosts:tableCosts]{tableCosts()}}}

  \item{CohortEconomics}{\code{\link[CohortEconomics:adjust_ps]{adjust_ps()}}, \code{\link[CohortEconomics:assess_balance]{assess_balance()}}, \code{\link[CohortEconomics:compile_trajectories]{compile_trajectories()}}, \code{\link[CohortEconomics:extract_hcru]{extract_hcru()}}, \code{\link[CohortEconomics:extractHcru]{extractHcru()}}, \code{\link[CohortEconomics:fit_ps]{fit_ps()}}, \code{\link[CohortEconomics:init]{init()}}, \code{\link[CohortEconomics:plot_ceac]{plot_ceac()}}, \code{\link[CohortEconomics:plot_plane]{plot_plane()}}, \code{\link[CohortEconomics:run_cea]{run_cea()}}, \code{\link[CohortEconomics:simulate_economics]{simulate_economics()}}, \code{\link[CohortEconomics:summarise_baseline]{summarise_baseline()}}, \code{\link[CohortEconomics:table_summary]{table_summary()}}}

  \item{CohortUtilisation}{\code{\link[CohortUtilisation:addEmergency]{addEmergency()}}, \code{\link[CohortUtilisation:addEmergencyCare]{addEmergencyCare()}}, \code{\link[CohortUtilisation:addEmergencyVisits]{addEmergencyVisits()}}, \code{\link[CohortUtilisation:addHospitalizations]{addHospitalizations()}}, \code{\link[CohortUtilisation:addInpatient]{addInpatient()}}, \code{\link[CohortUtilisation:addInpatients]{addInpatients()}}, \code{\link[CohortUtilisation:addOutpatientVisits]{addOutpatientVisits()}}, \code{\link[CohortUtilisation:addPrescriptions]{addPrescriptions()}}, \code{\link[CohortUtilisation:addProcedures]{addProcedures()}}, \code{\link[CohortUtilisation:addVisits]{addVisits()}}, \code{\link[CohortUtilisation:compute_hospitalization_cohorts]{compute_hospitalization_cohorts()}}, \code{\link[CohortUtilisation:computeHospitalizationCohorts]{computeHospitalizationCohorts()}}, \code{\link[CohortUtilisation:computeInfusionCohorts]{computeInfusionCohorts()}}, \code{\link[CohortUtilisation:plotUtilization]{plotUtilization()}}, \code{\link[CohortUtilisation:summariseUtilization]{summariseUtilization()}}, \code{\link[CohortUtilisation:tableUtilization]{tableUtilization()}}}
}}

