% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{samples,AnyHermesData-method}
\alias{samples,AnyHermesData-method}
\alias{samples}
\title{Sample IDs Accessor}
\usage{
\S4method{samples}{AnyHermesData}(object)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}
}
\value{
The character vector with the sample IDs.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}

Access the sample IDs, i.e. col names, of a \code{\link{AnyHermesData}} object with a
nicely named accessor method.
}
\examples{
a <- HermesData(summarized_experiment)
samples(a)
}
\seealso{
\code{\link[=genes]{genes()}} to access the gene IDs.
}
