% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{search_codes}
\alias{search_codes}
\title{get read codes}
\usage{
search_codes(
  regexes = list("heart\\\\s*failure", "ventric[ulare]+\\\\s*(dysfunction|failure)"),
  terms = list("Transplantation of heart and lung"),
  sources = c("icd9cm", "icd10cm", "icd10", "snomed", "opcs", "readcodes_v2",
    "readcodes_v3")
)
}
\arguments{
\item{regexes}{a list of regular expressions}

\item{terms}{a list of strings to search for}

\item{sources}{a vector of strings, one or more of c("icd9cm", "icd10cm", "icd10", "snomed", "opcs", "readcodes_v2", "readcodes_v3")}
}
\value{
a data.table
}
\description{
get read codes
}
