% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{show,HermesData-method}
\alias{show,HermesData-method}
\alias{show,RangedHermesData-method}
\title{Show Method for \code{HermesData} and \code{RangedHermesData} Objects}
\usage{
\S4method{show}{HermesData}(object)

\S4method{show}{RangedHermesData}(object)
}
\arguments{
\item{object}{(\code{HermesData}) or (\code{RangedHermesData}) \cr input.}
}
\description{
Show Method for \code{HermesData} and \code{RangedHermesData} Objects
}
\section{Methods (by class)}{
\itemize{
\item \code{HermesData}: A show method that displays additional information of \code{HermesData} objects.

\item \code{RangedHermesData}: A show method that displays additional information of \code{RangedHermesData} objects.
}}

\examples{
object <- HermesData(summarized_experiment)
show(object)
object <- HermesData(as(summarized_experiment, "RangedSummarizedExperiment"))
show(object)
}
