% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{show,HermesData-method}
\alias{show,HermesData-method}
\alias{show}
\alias{show,RangedHermesData-method}
\title{Show Method for \code{AnyHermesData} Objects}
\usage{
\S4method{show}{HermesData}(object)

\S4method{show}{RangedHermesData}(object)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr input.}
}
\description{
A show method that displays high-level information of \code{\link{AnyHermesData}} objects.
}
\note{
The same method is used for both \code{\link{HermesData}} and \code{\link{RangedHermesData}}
objects. We need to define this separately to have this method used instead of
the one inherited from \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}}.
}
\examples{
object <- HermesData(summarized_experiment)
object
}
