% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{subset}
\alias{subset}
\title{Subsetting HermesData Objects}
\value{
The subsetted \link{HermesData} object.
}
\description{
This method subsets HermesData objects, based on expressions involving the
\code{rowData} columns and the \code{colData} columns.
}
\note{
Note that this just inherits
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::subset,SummarizedExperiment-method()}}.
}
\examples{
a <- HermesData(summarized_experiment)
a
subset(a, subset = LowExpressionFlag, select = DISCSTUD == "N")

}
