% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\name{subset}
\alias{subset}
\title{Subsetting \code{AnyHermesData} Objects}
\arguments{
\item{x}{(\code{AnyHermesData})\cr object to subset from.}
}
\value{
The subsetted \code{\link{AnyHermesData}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This method subsets \code{\link{AnyHermesData}} objects, based on expressions involving the
\code{rowData} columns and the \code{colData} columns.
}
\note{
Note that this just inherits
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::subset,SummarizedExperiment-method()}}.
}
\examples{
a <- HermesData(summarized_experiment)
a
subset(a, subset = LowExpressionFlag, select = DISCSTUD == "N")
}
