% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{summarise_baseline}
\alias{summarise_baseline}
\title{Summarise baseline demographics and comorbidities}
\usage{
summarise_baseline(study)
}
\arguments{
\item{study}{A hermes_study object}
}
\value{
A hermes_hcru object with baseline_summary
}
\description{
Summarise baseline demographics and comorbidities
}
