% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summarized_experiment}
\alias{summarized_experiment}
\title{Example \code{SummarizedExperiment} Data}
\format{
A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} object with 20 samples covering
5085 features (\code{Entrez} gene IDs).
}
\source{
This is an artificial dataset designed to resemble real data.
}
\usage{
summarized_experiment
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This example \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} can be used to create a
\code{\link{HermesData}} object. It already contains the required columns in \code{rowData} and \code{colData}.
}
\seealso{
\code{\link{expression_set}} which contains similar data as a \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}}.
}
\keyword{datasets}
