% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\docType{class}
\name{HermesDataSummary-class}
\alias{HermesDataSummary-class}
\alias{.HermesDataSummary}
\alias{summary}
\alias{HermesDataSummary}
\alias{summary,AnyHermesData-method}
\alias{show,HermesDataSummary-method}
\title{Summary Method for HermesData Objects}
\usage{
\S4method{summary}{AnyHermesData}(object)

\S4method{show}{HermesDataSummary}(object)
}
\arguments{
\item{object}{(\code{HermesDataSummary}) \cr result from the summary method applied to
\link{AnyHermesData} object.}
}
\description{
Summary Method for HermesData Objects
}
\section{Methods (by class)}{
\itemize{
\item \code{AnyHermesData}: A summary method for \link{AnyHermesData} object that
creates a \link{HermesDataSummary} object.

\item \code{HermesDataSummary}: A show method prints summary description of \link{HermesDataSummary} object
generated by the \code{\link[=summary]{summary()}} method.
}}

\examples{
object <- HermesData(summarized_experiment)
object_summary <- summary(object)

# We can access parts of this S4 object with the slot operator.
str(object_summary)
slotNames(object_summary)
object_summary@lib_sizes
# Just calling the summary method like this will use the `show()` method.
summary(object)
}
