% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-methods.R
\docType{class}
\name{summary}
\alias{summary}
\alias{HermesDataSummary-class}
\alias{.HermesDataSummary}
\alias{HermesDataSummary}
\alias{summary,AnyHermesData-method}
\alias{show,HermesDataSummary-method}
\title{Summary Method for \code{AnyHermesData} Objects}
\usage{
summary(object, ...)

\S4method{summary}{AnyHermesData}(object)

\S4method{show}{HermesDataSummary}(object)
}
\arguments{
\item{object}{(\code{HermesDataSummary}) \cr result from the summary method applied to
\code{\link{AnyHermesData}} object.}

\item{...}{not used.}
}
\value{
An object of the corresponding summary class, here
\code{\link{HermesDataSummary}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Provides a concise summary of the content of \code{\link{AnyHermesData}} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{summary(AnyHermesData)}: A summary method for \code{\link{AnyHermesData}} object that
creates a \code{\link{HermesDataSummary}} object.

\item \code{show(HermesDataSummary)}: A show method prints summary description of \code{\link{HermesDataSummary}} object
generated by the \code{\link[=summary]{summary()}} method.

}}
\examples{
object <- hermes_data
object_summary <- summary(object)

# We can access parts of this S4 object with the `slot` function.
str(object_summary)
slotNames(object_summary)
slot(object_summary, "lib_sizes")

# Just calling the summary method like this will use the `show()` method.
summary(object)
}
