% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_library.R
\name{update_library}
\alias{update_library}
\title{Update code library}
\usage{
update_library(
  search_terms = c("heart failure", "cardiomyopathy", "myocardial infarction")
)
}
\arguments{
\item{search_terms}{a vector of strings to search the concept library phenotypes for}
}
\value{
boolean, success status
}
\description{
Update the internal code library from the Concept Library API
}
