% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermesData-validate.R
\name{validate}
\alias{validate}
\alias{validate_counts}
\alias{validate_cols}
\alias{all_na}
\alias{validate_non_empty}
\alias{validate_row_data}
\alias{validate_col_data}
\title{Internal Helper Functions for Validation of HermesData Objects}
\usage{
validate_counts(object)

validate_cols(required, actual)

all_na(x)

validate_non_empty(df)

validate_row_data(object)

validate_col_data(object)
}
\arguments{
\item{object}{(\code{SummarizedExperiment})\cr object to validate.}

\item{required}{(\code{character})\cr required column names.}

\item{actual}{(\code{actual})\cr actual column names.}

\item{x}{(\code{vector})\cr vector to check.}

\item{df}{(\code{data.frame})\cr data frame to validate.}
}
\value{
A character vector with the validation failure messages, or \code{NULL} in
case validation passes.
}
\description{
These functions are used internally only and therefore not exported. They work on
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} objects, and \link{HermesData} objects are
defined by successfully passing these validation checks.
}
\section{Functions}{
\itemize{
\item \code{validate_counts}: validates that the first assay is \code{counts} containing
non-missing, integer, non-negative values.

\item \code{validate_cols}: validates that required column names are contained in
actual column names.

\item \code{all_na}: checks whether the whole vector is \code{NA}.

\item \code{validate_non_empty}: validates that the data frame is not containing only
\code{NA}, per column.

\item \code{validate_row_data}: validates that the object contains \code{rowData} with
required columns.

\item \code{validate_col_data}: validates that the object contains \code{colData} with
required columns.
}}

